### 2018-05-03 #716 @ygrange
I'm not too familiar with the background application field (maybe I'm wrong and we could actually apply it to Astronomy; haven't really thought very deeply about it tbh) so I may mostly comment on the more technical details. Wouldn't mind having an in-depth look at it though. I will however be unavailable untill the end of next week.
