### 2018-05-03 e281f43
Job message signal/abort: poll to confirm
