### 2018-05-03 #716 @hjoliver
@arfon - the error says "could not find paper.bib", but that file is in the same location as paper.md in the repostory (which is "doc/joss-paper"). The pandoc command given in whedon docs works when I run it manually in that directory.  The instruction for authors does not say a specific location is required, or have I missed something?
