### 2018-05-04 7acce6e
Add get-host-metrics command & host selection logic, w/ tests
