+++ lib/cylc/remote.py
@@ -21,19 +21,25 @@ import os
 import sys
 import shlex
 import signal
+import unittest
+import json
+import socket
 from time import sleep
 from pipes import quote
 from posix import WIFSIGNALED
+from random import shuffle, choice
+from multiprocessing import Pool
 
 # CODACY ISSUE:
 #   Consider possible security implications associated with Popen module.
 # REASON IGNORED:
 #   Subprocess is needed, but we use it with security in mind.
-from subprocess import Popen, PIPE
+from subprocess import Popen, PIPE, CalledProcessError
 
 import cylc.flags
 from cylc.cfgspec.glbl_cfg import glbl_cfg
 from cylc.version import CYLC_VERSION
+from cylc.hostuserutil import get_fqdn_by_host
 
 
 def get_proc_ancestors():
@@ -62,91 +68,177 @@ def watch_and_kill(proc):
             break
 
 
-def remote_cylc_cmd(cmd, user=None, host=None, capture=False, manage=False,
-                    ssh_login_shell=None, ssh_cylc=None, stdin=None):
+def run_ssh_cmd(command, stdin=None, capture_process=False,
+                capture_status=False, manage=False):
     """Run a given cylc command on another account and/or host.
 
     Arguments:
-        cmd (list): command to run remotely.
-        user (string): user ID for the remote login.
-        host (string): remote host name. Use 'localhost' if not specified.
-        capture (boolean):
+        command (list):
+            command inclusive of all opts and args required to run via ssh.
+        stdin (file):
+            If specified, it should be a readable file object.
+            If None, `open(os.devnull)` is set if output is to be captured.
+        capture_process (boolean):
             If True, set stdout=PIPE and return the Popen object.
+        capture_status (boolean):
+            If True, and the remote command is unsuccessful, return the
+            associated exit code instead of exiting with an error.
         manage (boolean):
             If True, watch ancestor processes and kill command if they change
             (e.g. kill tail-follow commands when parent ssh connection dies).
-        ssh_login_shell (boolean):
-            If True, launch remote command with `bash -l -c 'exec "$0" "$@"'`.
-        ssh_cylc (string):
-            Location of the remote cylc executable.
-        stdin (file):
-            If specified, it should be a readable file object.
-            If None, it will be set to `open(os.devnull)` and the `-n` option
-            will be added to the SSH command line.
 
     Return:
-        If capture=True, return the Popen object if created successfully.
-        Otherwise, return the exit code of the remote command.
-    """
-    if host is None:
-        host = "localhost"
-    if user is None:
-        user_at_host = host
-    else:
-        user_at_host = '%s@%s' % (user, host)
+        * If capture_process=True, the Popen object if created successfully.
+        * Else True if the remote command is executed successfully, or
+          if unsuccessful and capture_status=True the remote command exit code.
+        * Otherwise exit with an error message."""
 
-    # Build the remote command
-    command = shlex.split(
-        str(glbl_cfg().get_host_item('ssh command', host, user)))
-    if stdin is None:
-        command.append('-n')
-        stdin = open(os.devnull)
-    command.append(user_at_host)
-
-    # Pass cylc version through.
-    command += ['env', r'CYLC_VERSION=%s' % CYLC_VERSION]
-
-    if ssh_login_shell is None:
-        ssh_login_shell = glbl_cfg().get_host_item(
-            'use login shell', host, user)
-    if ssh_login_shell:
-        # A login shell will always source /etc/profile and the user's bash
-        # profile file. To avoid having to quote the entire remote command
[…]
