### 2018-05-04 #716 @ygrange
So in principle I'd happily do it but I will be looking at it from the direction of a different field (so I will judge the documentation on how clear it is for me as an outsider for example). I leave it to you to judge if that's an angle you want to have covered (not sure of joss allows multiple reviewers to cover different angles).

Going offline for a week now and I'll see what's expected from me next week :)
