### 2018-05-04 #716 @hjoliver
Thanks @ygrange.  Regarding documentation, I don't think you'll find it too weather-specialized, except perhaps for the introductory section of the User Guide which dates back to the inception of the project. I'm not sure if this is good news or bad for reviewers, but Cylc is a reasonably large, mature application, so the User Guide is quite big.  I suggest you run through one of the presentations here https://cylc.github.io/cylc/documentation.html first as a quick overview.
