### 2018-05-08 8f31612
Fix suicide trigger & logic
Also flatten single item in "(...)" in graph.
