### 2018-05-09 cac15ec
remove duplicate parsers
