### 2018-05-10 7130d22
Kill proc kill commands that take too long.
