### 2018-05-11 3e4ee3a
Changes to allow grouped sub-graphs to be visible on GCylc startup.
