+++ lib/cylc/gui/updater_graph.py
@@ -33,6 +33,7 @@ from cylc.mkdir_p import mkdir_p
 from cylc.network.httpclient import ClientError
 from cylc.task_id import TaskID
 from cylc.task_state import TASK_STATUS_RUNAHEAD
+from cylc.cfgspec.gcylc import gcfg
 
 
 def compare_dict_of_dict(one, two):
@@ -81,7 +82,8 @@ class GraphUpdater(threading.Thread):
         self.best_fit = True  # zoom to page size
         self.normal_fit = False  # zoom to 1.0 scale
         self.crop = False
-        self.subgraphs_on = False   # organise by cycle point.
+        # organise by cycle point.
+        self.subgraphs_on = gcfg.get(["sub-graphs on"])
 
         self.descendants = {}
         self.all_families = []
