+++ INSTALL.md
@@ -22,9 +22,9 @@ as `/opt`:
 
 ```bash
 cd /opt
-tar xzf cylc-7.4.0.tar.gz
+tar xzf cylc-7.7.0.tar.gz
 # DO NOT CHANGE THE NAME OF THE UNPACKED CYLC SOURCE DIRECTORY.
-cd cylc-7.4.0
+cd cylc-7.7.0
 export PATH=$PWD/bin:$PATH
 make
 ```
@@ -41,13 +41,13 @@ long-running suites (and their task jobs) to stick with older versions if
 necessary. The wrapper should be edited to point to the Cylc install location:
 
 ```bash
-cp /opt/cylc-7.4.0/admin/cylc-wrapper /usr/local/bin/cylc
+cp /opt/cylc-7.7.0/sbin/cylc-wrapper /usr/local/bin/cylc
 # (now edit '/usr/local/bin/cylc' as per in-file instructions...)
 ```
 
 Finally, make a symlink to the latest installed version:
 ```bash
-ln -s /opt/cylc-7.4.0 /opt/cylc
+ln -s /opt/cylc-7.7.0 /opt/cylc
 ```
 (This will be the default version invoked by the wrapper if a specific version is not requested via \lstinline=$CYLC_VERSION=.
 
