+++ doc/src/cylc-user-guide/cug.tex
@@ -477,7 +477,7 @@ separately.
 
 Cylc releases can be downloaded from from \url{https://cylc.github.io/cylc}.
 
-The wrapper script \lstinline=admin/cylc-wrapper= should be installed as
+The wrapper script \lstinline=sbin/cylc-wrapper= should be installed as
 \lstinline=cylc= in the system executable search path (e.g.\
 \lstinline=/usr/local/bin/=) and modified slightly to point to a location
 such as \lstinline=/opt= where successive Cylc releases will be unpacked
