### 2018-05-11 f3902ab
Update docs a bit.
