### 2018-05-12 7961bd5
Updated changes.md for 7.7.0 [skip ci]
