+++ doc/src/cylc-user-guide/job-host.tex
@@ -2,6 +2,9 @@
 
 \lstset{language=jobhosts}
 
+{\em This section is slightly out of date: all Cylc (but not Rose) commands
+now act via \lstinline=cylc= sub-commands on remote job hosts.}
+
 This appendix shows transcripts of all \lstinline=ssh=, \lstinline=scp=, and
 \lstinline=rsync= invocations from a Cylc suite called \lstinline=test= started by
 \lstinline=rose suite-run= on \lstinline=vagrant@localhost=. The suite
