+++ INSTALL.md
@@ -1,17 +1,14 @@
 # Cylc Installation.
 
-**See [The Cylc User Guide](https://cylc.github.io/cylc/documentation.html) for
-detailed instructions.**
+**See also [The Cylc User Guide](https://cylc.github.io/cylc/documentation.html).**
 
 Cylc must be installed on suite and task job hosts. Third-party dependencies
 (below) are not required on job hosts.
 
-### Required Third-party Software Packages
+### Third-party Software Packages
 
-These can be installed on suite hosts, and updated very infrequently.
-
- * graphviz
- * pygraphviz
+Install the packages listed in the **Installation** section of the User Guide.
+See also *Check Software Installation* below.
 
 ### Installing Cylc
 
@@ -37,11 +34,9 @@ When you type `make`:
   * The Cylc documentation is generated from source and put in doc/install/ (if
     you have pdflatex, tex4ht, and several other LateX packages installed).
 
-Cylc commands should be invoked via a central wrapper script that you need to
-copy from sbin/cylc-wrapper to (say) /usr/local/bin/ and rename it as "cylc".
-The wrapper selects between available versions, allowing long-running suites
-(and their task jobs) to stick with older versions if necessary. The wrapper
-should be edited to point to the Cylc install location:
+Now copy the wrapper script `sbin/cylc-wrapper` to (say) `/usr/local/bin/`,
+rename it as just `cylc`, and edit it - as per instructions in the file - to
+point to the Cylc install location:
 
 ```bash
 cp /opt/cylc-7.7.0/sbin/cylc-wrapper /usr/local/bin/cylc
@@ -52,7 +47,54 @@ Finally, make a symlink to the latest installed version:
 ```bash
 ln -s /opt/cylc-7.7.0 /opt/cylc
 ```
-(The wrapper will invoke this version if \lstinline=$CYLC_VERSION= is not set).
+The central wrapper will invoke `cylc-$CYLC_VERSION` if `$CYLC_VERSION` is set
+and that version is installed, or else the symlinked version as default. (Or
+developers can set `$CYLC_HOME` to point to their local Cylc clone). Suite
+server programs set `$CYLC_VERSION` to ensure that client programs invoked by
+running task jobs (for messaging etc.) run from the same version as the
+server.
+
+### Check Software Installation
+
+```
+$ cylc check-software
+Checking your software...
+
+Individual results:
+===============================================================================
+Package (version requirements)                          Outcome (version found)
+===============================================================================
+                              *REQUIRED SOFTWARE*                              
+Python (2.6+, <3).....................FOUND & min. version MET (2.7.12.final.0)
+
+       *OPTIONAL SOFTWARE for the GUI & dependency graph visualisation*       
+Python:pygtk (2.0+)...........................FOUND & min. version MET (2.24.0)
+graphviz (any)...................................................FOUND (2.38.0)
+Python:pygraphviz (any)...........................................FOUND (1.3.1)
+
+                  *OPTIONAL SOFTWARE for the HTML User Guide*                  
+ImageMagick (any)...............................................FOUND (6.8.9-9)
+
+            *OPTIONAL SOFTWARE for the HTTPS communications layer*            
+Python:urllib3 (any).............................................FOUND (1.13.1)
+Python:OpenSSL (any).............................................FOUND (17.2.0)
+Python:requests (2.4.2+).......................FOUND & min. version MET (2.9.1)
+
+                 *OPTIONAL SOFTWARE for the LaTeX User Guide*                 
+TeX:framed (any)....................................................FOUND (n/a)
+TeX (3.0+)................................FOUND & min. version MET (3.14159265)
+TeX:preprint (any)..................................................FOUND (n/a)
+TeX:tex4ht (any)....................................................FOUND (n/a)
+TeX:tocloft (any)...................................................FOUND (n/a)
[…]
