+++ doc/src/cylc-user-guide/cug.tex
@@ -383,7 +383,7 @@ previous section, emerges naturally at run time.
 
 Cylc runs on Unix variants, usually Linux, and including Apple OS X.
 
-\subsection{External Software Packages}
+\subsection{Third-Party Software Packages}
 
 {\bf Python \lstinline@>=@ 2.6} is required (but not yet Python 3). Python
 should already be installed in your Linux system. \url{https://python.org}.
