+++ CHANGES.md
@@ -1,7 +1,96 @@
 # Selected Cylc Changes
 
-For the full list of all changes for each release see [closed
-milestones](https://github.com/cylc/cylc/milestones?state=closed).
+Internal changes that do not directly affect users may not be listed here.  For
+all changes see the [closed
+milestones](https://github.com/cylc/cylc/milestones?state=closed) for each
+release.
+
+-------------------------------------------------------------------------------
+## __cylc-7.7.0 (2018-05-12)__
+
+### Enhancements
+
+[#2582](https://github.com/cylc/cylc/pull/2582) - improve client/server
+interface, including: `cylc message` can send multiple messages at once, with
+different severities; server ignores messages from superseded job submits;
+running jobs detect that the suite has been cold-started under them and will
+not attempt to connect; ssh-based indirect client-server communication now
+works automatically for all clients, not just messaging.
+
+[#2624](https://github.com/cylc/cylc/pull/2624) - rewrite `cylc cat-log`
+for enhanced functionality and some bug fixes.
+
+[#2582](https://github.com/cylc/cylc/pull/2582), 
+[#2624](https://github.com/cylc/cylc/pull/2624), and earlier changes: all job
+host actions are now done by remote `cylc` subcommands that are compatible with
+ssh whitelisting.
+
+[#2590](https://github.com/cylc/cylc/pull/2590) - replaced the fixed process
+pool with direct management of individual sub-processes (for external commands
+executed by the server program).
+
+[#2561](https://github.com/cylc/cylc/pull/2561) - pass resolved triggering
+dependencies to `$CYLC_TASK_DEPENDENCIES` in job environments.
+
+[#2639](https://github.com/cylc/cylc/pull/2639) - date-time cycling: document
+availability of 365-day (never a leap year) and 366-day (always a leap year)
+calendars.
+
+[#2597](https://github.com/cylc/cylc/pull/2597) - emit a "late" event if a task
+has not triggered by a user-defined real-time offset relative to cycle point. 
+
+[#2648](https://github.com/cylc/cylc/pull/2648) - improve version reporting.
+Note than non-standard lowercase `cylc -v` is now gone; use `cylc -V` or `cylc
+--version`, with optional `--long` format to print as well as version.
+
+[#2620](https://github.com/cylc/cylc/pull/2620) - re-document the long-dated
+`[special tasks]sequential` and recommend using explicit dependencies in the
+graph instead.
+
+[#2584](https://github.com/cylc/cylc/pull/2584) - internal queues now release
+task jobs on a FIFO (First In, First Out) basis, rather than randomly.
+
+[#2538](https://github.com/cylc/cylc/pull/2538) - remove leading whitespace
+from multi-line `script` items in task definitions, for cleaner job scripts.
+
+### Fixes
+
+[#2593](https://github.com/cylc/cylc/pull/2593) - fix polling after job
+execution timeout (configured pre-poll delays were being ignored).
+
+[#2656](https://github.com/cylc/cylc/pull/2656) - fix suicide triggers with
+multiple prerequisites in the same graph line. 
+
+[#2638](https://github.com/cylc/cylc/pull/2618) - fix duplicate "failed" task
+events after `cylc stop --kill`.
+
+[#2653](https://github.com/cylc/cylc/pull/2653) - tidy and correct the main `cylc
+help` documentation.
+
+[#2644](https://github.com/cylc/cylc/pull/2644), 
+[#2646](https://github.com/cylc/cylc/pull/2646) - fix some graph parsing edge
+cases.
+
[…]
