+++ doc/src/cylc-user-guide/cug-html.tex
@@ -11,7 +11,8 @@
 \usepackage{textcomp}  % \textgreater, \textless
 \usepackage{framed}
 \usepackage{listings}
-  \usepackage{courier}
+\usepackage{courier}
+\usepackage{dirtree}
 \usepackage{amsmath}
 \usepackage{verbatim}
 \usepackage{graphicx}             % latex, eps
