+++ doc/src/cylc-user-guide/cug.tex
@@ -425,6 +425,7 @@ required may vary by distribution, e.g.:
     \item texlive-framed
     \item texlive-preprint (for \lstinline=fullpage.sty=)
     \item texlive-tex4ht
+    \item texlive-generic-extra (for \lstinline=dirtree.sty=)
 \end{myitemize}
 
 To generate the HTML User Guide {\bf ImageMagick} is also needed.
