+++ doc/src/cylc-user-guide/cug.tex
@@ -508,11 +508,10 @@ separately.
 
 Cylc releases can be downloaded from from \url{https://cylc.github.io/cylc}.
 
-The wrapper script \lstinline=sbin/cylc-wrapper= should be installed as
-\lstinline=cylc= in the system executable search path (e.g.\
-\lstinline=/usr/local/bin/=) and modified slightly to point to a location
-such as \lstinline=/opt= where successive Cylc releases will be unpacked
-side by side.
+The wrapper script \lstinline=usr/bin/cylc= should be installed to
+the system executable search path (e.g.\ \lstinline=/usr/local/bin/=) and
+modified slightly to point to a location such as \lstinline=/opt= where
+successive Cylc releases will be unpacked side by side.
 
 To install Cylc, unpack the release tarball in the right location, e.g.\
 \lstinline=/opt/cylc-7.7.0=, type \lstinline=make= inside the release
