### 2018-05-13 a310949
Moved examples into etc dir.
