+++ doc/src/cylc-user-guide/cug.tex
@@ -851,7 +851,7 @@ $ cylc import-examples /tmp
 
 \subsection{Rename The Imported Tutorial Suites}
 
-Suites can be renames by simply renaming (i.e.\ moving) their run directories.
+Suites can be renamed by simply renaming (i.e.\ moving) their run directories.
 Make the tutorial suite names shorter, and print their locations with
 \lstinline=cylc print=:
 
@@ -892,7 +892,7 @@ $ echo $?
 Here's the traditional {\em Hello World} program rendered as a cylc
 suite:
 \lstset{language=suiterc}
-\lstinputlisting{../../../examples/tutorial/oneoff/basic/suite.rc}
+\lstinputlisting{../../../etc/examples/tutorial/oneoff/basic/suite.rc}
 \lstset{language=transcript}
 
 Cylc suites feature a clean separation of scheduling configuration,
@@ -913,7 +913,7 @@ prototyping of new suites, it is submitted to run as a background job on
 the suite host. In fact cylc even provides a default task implementation
 that makes the entire \lstinline=[runtime]= section technically optional:
 \lstset{language=suiterc}
-\lstinputlisting{../../../examples/tutorial/oneoff/minimal/suite.rc}
+\lstinputlisting{../../../etc/examples/tutorial/oneoff/minimal/suite.rc}
 \lstset{language=transcript}
 (the resulting {\em dummy task} just prints out some identifying
 information and exits).
@@ -1420,7 +1420,7 @@ represents a task, or if it is inherited by other namespaces, a {\em
 family}. This allows common configuration to be factored out of related
 tasks very efficiently.
 \lstset{language=suiterc}
-\lstinputlisting{../../../examples/tutorial/oneoff/inherit/suite.rc}
+\lstinputlisting{../../../etc/examples/tutorial/oneoff/inherit/suite.rc}
 The \lstinline=[root]= namespace provides defaults for all tasks in the suite.
 Here both tasks inherit \lstinline=script= from \lstinline=root=, which they
 customize with different values of the environment variable
@@ -1459,7 +1459,7 @@ i.e.\ if \lstinline=foo= succeeds, trigger all members of
 \lstinline=GREETERS= at once. Here's the full suite with runtime
 hierarchy shown:
 \lstset{language=suiterc}
-\lstinputlisting{../../../examples/tutorial/oneoff/ftrigger1/suite.rc}
+\lstinputlisting{../../../etc/examples/tutorial/oneoff/ftrigger1/suite.rc}
 
 (Note that we recommend given ALL-CAPS names to task families to help
 distinguish them from task names. However, this is just a convention).
@@ -1591,11 +1591,11 @@ basic Hello World suite and cut the implementation of the task
 \lstinline=hello= out to a file \lstinline=hello.sh= in the suite
 bin directory:
 \lstset{language=bash}
-\lstinputlisting{../../../examples/tutorial/oneoff/external/bin/hello.sh}
+\lstinputlisting{../../../etc/examples/tutorial/oneoff/external/bin/hello.sh}
 Make the task script executable, and change the \lstinline=hello= task
 runtime section to invoke it:
 \lstset{language=suiterc}
-\lstinputlisting{../../../examples/tutorial/oneoff/external/suite.rc}
+\lstinputlisting{../../../etc/examples/tutorial/oneoff/external/suite.rc}
 
 If you run the suite now the new greeting from the external task script
 should appear in the \lstinline=hello= task stdout log. This works
@@ -1632,7 +1632,7 @@ belonging to that sequence may only run once.
 
 Open the \lstinline=tut/cycling/one= suite:
[…]
