### 2018-05-13 86cb49e
Moved syntax files to etc dir.
