+++ doc/src/cylc-user-guide/cug.tex
@@ -2179,10 +2179,10 @@ Jinja2 (\ref{Jinja2}) also has template inclusion functionality.
 Cylc comes with syntax files for a number of text editors:
 \lstset{language=transcript}
 \begin{lstlisting}
-$CYLC_DIR/conf/cylc.vim     # vim
-$CYLC_DIR/conf/cylc-mode.el # emacs
-$CYLC_DIR/conf/cylc.lang    # gedit (and other gtksourceview programs)
-$CYLC_DIR/conf/cylc.xml     # kate
+$CYLC_DIR/etc/syntax/cylc.vim     # vim
+$CYLC_DIR/etc/syntax/cylc-mode.el # emacs
+$CYLC_DIR/etc/syntax/cylc.lang    # gedit (and other gtksourceview programs)
+$CYLC_DIR/etc/syntax/cylc.xml     # kate
 \end{lstlisting}
 Refer to comments at the top of each file to see how to use them.
 
