### 2018-05-13 aba9c8e
Move remain source files in conf to etc.
