+++ doc/src/cylc-user-guide/gcylcrc.tex
@@ -187,7 +187,7 @@ using the tool bar.
 Set the task state color theme, common to all views, to use
 initially. The color theme can be changed later using the tool bar.
 See \lstinline@gcylc.rc.eg@ and \lstinline@themes.rc@ in
-\lstinline@$CYLC_DIR/conf/gcylcrc/@ for how to modify existing color
+\lstinline@$CYLC_DIR/etc/gcylcrc/@ for how to modify existing color
 themes or define your own. Use \lstinline@cylc get-gui-config@
 to list your available themes.
 
@@ -246,7 +246,7 @@ For the attribute values, COLOR and FONTCOLOR can be color names from the X11
 rgb.txt file, e.g.\ \lstinline=SteelBlue=; or hexadecimal color codes, e.g.\ 
 \lstinline@#ff0000@ for red; and STYLE can be ``filled'' or ``unfilled''.
 See \lstinline@gcylc.rc.eg@ and \lstinline@themes.rc@ in
-\lstinline@$CYLC_DIR/conf/gcylcrc/@ for examples.
+\lstinline@$CYLC_DIR/etc/gcylcrc/@ for examples.
 
 \paragraph[STATE]{[themes] \textrightarrow [[THEME]] \textrightarrow STATE}
 
@@ -265,4 +265,4 @@ For the attribute values, COLOR and FONTCOLOR can be color names from the X11
 rgb.txt file, e.g.\ \lstinline=SteelBlue=; or hexadecimal color codes, e.g.\ 
 \lstinline@#ff0000@ for red; and STYLE can be ``filled'' or ``unfilled''.
 See \lstinline@gcylc.rc.eg@ and \lstinline@themes.rc@ in
-\lstinline@$CYLC_DIR/conf/gcylcrc/@ for examples.
+\lstinline@$CYLC_DIR/etc/gcylcrc/@ for examples.
