### 2018-05-13 3ef8150
Updated tutorial suite tests.
