### 2018-05-14 2fea42f
Finish off external command timeout.
* invoke callback after kill
* global config timeout, PT10M default
