### 2018-05-14 #716 @arfon
> From the list of potential reviewers `@ygrange` looks good (scientific workflow management, Python).

@hjoliver - I'd like to get an extra reviewer for this submission. Could you take a look at this list and suggest some others? https://bit.ly/joss-reviewers

Also, if you know the GitHub handles of any other potential good reviewers feel free to mention them here.
