### 2018-05-14 3856ed3
Added external command timeout tests.
