### 2018-05-14 be2f62c
Documented external command timeout.
