### 2018-05-14 1201e61
Address feedback: trivial comment update [skip ci].
