### 2018-05-14 3e39279
Final update to change log [skip ci].
