### 2018-05-14 abdfaaf
Corrected "cylc message" help and docs.
