### 2018-05-15 c039207
Moved global config to etc.
