+++ doc/src/cylc-user-guide/siterc.tex
@@ -808,7 +808,7 @@ Global site/user defaults for~\ref{TaskEventHandling}.
 \subsection{[test battery]}
 
 Settings for the automated development tests. Note the test battery reads
-\lstinline=$CYLC_DIR/conf/global-tests.rc= instead of the normal site/user
+\lstinline=$CYLC_DIR/etc/global-tests.rc= instead of the normal site/user
 global config files.
 
 \subsubsection[remote host with shared fs]{[test battery] \textrightarrow remote host with shared fs}
