### 2018-05-15 d0c77b8
Changed $CYLC_DIR to <cylc-dir> in docs.
The shell variable would not normally be defined in user environments.
