+++ doc/src/cylc-user-guide/cug.tex
@@ -676,41 +676,26 @@ A simpler six-hourly cycling workflow should be considered first, however.
 \label{SiteAndUserConfiguration}
 
 Cylc site and user global configuration files contain settings that affect all
-suites. Some of these, such as the range of network ports used by cylc,
-should be set at site level,
+suites.  Some of these, such as the range of network ports used by cylc,
+should be set at site level. Legal items, values, and system defaults are
+documented in (\ref{SiteRCReference}).
+
 \lstset{language=transcript}
 \begin{lstlisting}
 # cylc site global config file
-/path/to/cylc/conf/global.rc
-# Deprecated path to cylc site global config file
-/path/to/cylc/conf/siterc/site.rc
+<cylc-dir>/etc/global.rc
 \end{lstlisting}
 Others, such as the preferred text editor for suite definitions,
 can be overridden by users,
 \lstset{language=transcript}
 \begin{lstlisting}
 # cylc user global config file
-~/.cylc/global.rc
-# Deprecated cylc user global config file
-~/.cylc/user.rc
+~/.cylc/$(cylc --version)/global.rc  # e.g. ~/.cylc/7.7.0/global.rc
 \end{lstlisting}
 
-The \lstinline=cylc get-site-config= command retrieves current
-global settings consisting of cylc defaults overridden by site settings,
-if any, overridden by user settings, if any.
-If you need to generate an example user global config file filled with the
-default values commented out (useful as a quick reference), you can do:
-\lstset{language=transcript}
-\begin{lstlisting}
-$ cylc get-site-config | sed 's/^/#/' > ~/.cylc/global.rc.example
-\end{lstlisting}
-Settings that do not need to be changed should be deleted to reduce conflicts
-with site changes and changes in future versions of cylc.
-Copy the example file into \lstinline=~/.cylc/global.rc= if you need to apply
-any changes.
-
-Legal items, values, and system defaults are documented in
-(\ref{SiteRCReference}).
+The file \lstinline=<cylc-dir>/etc/global.rc.eg= contains instructions on how
+to generate and install site and user global config files:
+\lstinputlisting{../../../etc/global.rc.eg}
 
 %\pagebreak
 \section{Tutorial}
@@ -727,7 +712,7 @@ cylc on suite definitions:
 
 \lstset{language=suiterc}
 \begin{lstlisting}
-# $HOME/.cylc/global.rc
+# $HOME/.cylc/$(cylc --version)/global.rc
 [editors]
     terminal = vim
     gui = gvim -f
@@ -2179,10 +2164,10 @@ Jinja2 (\ref{Jinja2}) also has template inclusion functionality.
 Cylc comes with syntax files for a number of text editors:
 \lstset{language=transcript}
 \begin{lstlisting}
-$CYLC_DIR/etc/syntax/cylc.vim     # vim
-$CYLC_DIR/etc/syntax/cylc-mode.el # emacs
-$CYLC_DIR/etc/syntax/cylc.lang    # gedit (and other gtksourceview programs)
-$CYLC_DIR/etc/syntax/cylc.xml     # kate
+<cylc-dir>/etc/syntax/cylc.vim     # vim
+<cylc-dir>/etc/syntax/cylc-mode.el # emacs
+<cylc-dir>/etc/syntax/cylc.lang    # gedit (and other gtksourceview programs)
+<cylc-dir>/etc/syntax/cylc.xml     # kate
 \end{lstlisting}
 Refer to comments at the top of each file to see how to use them.
 
@@ -3036,7 +3021,7 @@ where suite initial and final cycle points can be assumed. Some examples:
 \subparagraph{Example}
 
 The tutorial illustrates integer cycling in~\ref{TutInteger}, and
[…]
