+++ CHANGES.md
@@ -18,6 +18,8 @@ release.
      the old `conf/` location.
    - look for user `global.rc` first in `~/.cylc/$CYLC_VERSION/`, then (if not
      found) in the old `~/.cylc/` location. See notes in `/etc/global.rc.eg`.
+   - *note the very old `site.rc` and `user.rc` global config filename variants
+     are now obsolete and will not be recognized*
    - moved central cylc wrapper template to `usr/bin/cylc`
    - moved editor syntax files to `etc/syntax/`
    - moved Cylc bash completion file to `etc/cylc-bash-completion`
