### 2018-05-15 31cb150
Fix tutorial ref tests.
(These were regenerated a few commits ago without updated the generating
script with --no-detach)
