### 2018-05-15 #716 @hjoliver
@arfon - from the list I could suggest @sarats and @Chilipp, who seem to be involved in Earth System Modeling, one of Cylc's primary use cases currently.

If they don't work out I can try to find others, but unfortunately most of my immediate contacts in the subject area are already Cylc users or developers (which is presumably not OK?)
