### 2018-05-15 #716 @Chilipp
@arfon and @hjoliver - Looks interesting, I would be willing to review it
