### 2018-05-16 #737 @whedon
[REVIEW]: Cylc: a Workflow Engine for Automating Cycling Systems
**Submitting author:** @hjoliver (<a href="http://orcid.org/0000-0002-5715-5279">Hilary Oliver</a>)
**Repository:** <a href="https://github.com/cylc/cylc" target ="_blank">https://github.com/cylc/cylc</a>
**Version:** cylc-7.7.0

- [x] **Version:** Does the release version given match the GitHub release (cylc-7.7.0)?

- [x] **Version:** Does the release version given match the GitHub release (cylc-7.6.1)?
