### 2018-05-16 #716 @hjoliver
@sarats, @Chilipp - many thanks from me too.
