+++ index.md
@@ -9,8 +9,10 @@ indefinitely.
 
 There are several reasons why tasks might need to be cycled:
 
- * In real time environmental forecasting systems, new forecasts are initiated
-   at regular intervals as new driving data comes in.
+ * To run successive cycles of a real time environmental forecasting system
+   (where in real time operation new forecasts are initiated at regular
+   intervals as new driving data comes in; but in catch-up or historical-run
+   mode the system dependencies may allow multiple cycles to run concurrently)
 
  * To split long model runs into a sequence of smaller runs, with associated
    processing tasks for each chunk.
