### 2018-05-16 4aa1069
Move job-init-env.sh to etc/.
