+++ doc/src/cylc-user-guide/scripts/make-pdf.sh
@@ -27,8 +27,7 @@ cp -r *.tex cylc-version.txt titlepic.sty $DEST
 
 cd $DEST
 ls *.tex | xargs -n 1 perl -pi -e 's@graphics/png/orig@../graphics/png/orig@g'
-ls *.tex | xargs -n 1 perl -pi -e 's@\.\./etc/examples/@../../etc/examples/@g'
-ls *.tex | xargs -n 1 perl -pi -e 's@\.\./conf/@../../conf/@g'
+ls *.tex | xargs -n 1 perl -pi -e 's@\.\./etc/@../../etc/@g'
 perl -pi -e 's@categories/@../categories/@g' commands.tex
 perl -pi -e 's@commands/@../commands/@g' commands.tex
 perl -pi -e 's@cylc.txt@../cylc.txt@g' commands.tex
