### 2018-05-16 53c660e
scan/gscan: consistent default behaviour
By default, scan only running suites of current user by looking at
`~/cylc-run/*/.service/contact` of current user.
