### 2018-05-16 66bffb0
cylc scan: Update CHANGES.md + on daemonize help
Note that `cylc scan --name=PATTERN` will require a pattern to match
full names of suites.
