### 2018-05-17 c46b3aa
cat-log suicide only if invoked remotely.
