### 2018-05-19 7d7a240
cat-log: fix remote tail tidy, properly.
