### 2018-05-20 7e7a29f
cat-log: disallow remote edit.
