### 2018-05-21 5c2cdb0
trap error after cylc message success
