### 2018-05-21 676670c
remove unnecessary for loops
