### 2018-05-22 eb05784
Omit half of job-host section from HTML CUG.
(Bad HTML formatting.)
