+++ doc/src/cylc-user-guide/cug.tex
@@ -7667,7 +7667,8 @@ cylc. It is available under the open source GPL license.
 
 \pagebreak
 
-\input{job-host.tex}
+\input{job-host-1.tex}
+\input{job-host-2.tex}
 
 \pagebreak
 
