+++ etc/dev-bin/defn-order-test.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python2
 
 import time, string, random
 from copy import deepcopy
