+++ lib/parsec/tests/getcfg/bin/one-line.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python2
 """Check that single-line config print works"""
 
 import os, sys
