+++ lib/cylc/suite_srv_files_mgr.py
@@ -526,7 +526,9 @@ To start a new run, stop the old one first with one or more of these:
         cert_obj.gmtime_adj_notAfter(10 * 365 * 24 * 60 * 60)  # 10 years.
         cert_obj.set_issuer(cert_obj.get_subject())
         cert_obj.set_pubkey(pkey_obj)
-        cert_obj.set_serial_number(1)
+        # Set a random serial number to avoid the browser error
+        # SEC_ERROR_REUSED_ISSUER_AND_SERIAL.
+        cert_obj.set_serial_number(int(os.urandom(20).encode('hex'), 16))
         cert_obj.add_extensions([ext])
         cert_obj.sign(pkey_obj, 'sha256')
         self._dump_item(
