### 2018-05-24 0af139c
Fix restart task action timers load old poll timer
It may accidentally load the data structure as an event timer, which can
cause a strange error.
