### 2018-05-24 96c0daf
gui log viewer: undo bold font
