### 2018-05-27 6cdbafa
Use uuid4 to get ssl cert serial number.
