### 2018-05-29 cf98344
Add Jinja2 filter for ISO8601 duration -> seconds
Map an ISO8601 duration to a floating-point number
of seconds for use in suite definition files that
require a duration in non-ISO8601 units.
