### 2018-05-29 9eaac78
Add Jinja2 filter for ISO8601 duration -> hours
Map an ISO8601 duration to a floating-point number
of hours for use in suite definition files that
require a duration in non-ISO8601 format.
