### 2018-05-29 1684355
Forgot to update variable names - tested to work.
