### 2018-05-29 febd297
Minor updates to JOSS paper.
