+++ doc/joss-paper/paper.bib
@@ -1,9 +1,9 @@
 @misc{zenodo,
   publisher = {Zenodo},
   author = {Hilary Oliver and Matt Shin and Oliver Sanders et.al.},
-  doi = {10.5281/zenodo.1208732},
+  doi = {10.5281/zenodo.1251084},
   year = {2018},
-  month = {Mar},
-  day = {27},
-  title = {cylc/cylc: cylc-7.6.1},
+  month = {May},
+  day = {22},
+  title = {cylc/cylc: cylc-7.7.0},
 }
