+++ doc/joss-paper/paper.md
@@ -24,6 +24,8 @@ tags:
 - ocean
 - Python
 - workflow
+- scheduling
+- automation
 affiliations:
 - index: 1
   name: National Institute of Water and Atmospheric Research (NIWA), New Zealand
@@ -35,19 +37,19 @@ affiliations:
 
 # Summary
 
-Cylc is a workflow engine for orchestrating complex *suites* of inter-dependent
-distributed cycling (repeating) tasks, as well as ordinary non-cycling
+Cylc is a workflow engine for orchestrating complex distributed *suites* of
+inter-dependent cycling (repeating) tasks, as well as ordinary non-cycling
 workflows. It has been widely adopted for weather, climate, and related
 forecasting applications in research and production HPC environments, and it is
 now part of the official software infrastructure for the Unified Model
 atmospheric model. Cylc is written in Python and developed primarily by NIWA
 (NZ) and Met Office (UK). It has strong support for large production systems,
 but ease of use for individuals with smaller workflow automation requirements
-remains a key priority, and despite its core user base it is not in any way
-specialized to environmental forecasting.
+remains a key priority. Despite its origins, Cylc is not in any way specialized
+to environmental forecasting.
 
 In cycling workflows tasks repeat on sequences that may represent forecast
-cycles, chunks of a simulation that is too long for a single run, steps in some
+cycles, chunks of a simulation too long for a single run, steps in some
 multi-program iterative process (e.g. for optimizing model parameters), or
 datasets to be processed as they are generated or received, and so forth.
 Cycling in Cylc is controlled by ISO 8601 date-time recurrence expressions
@@ -57,13 +59,13 @@ never-ending, workflows (rather than simply a succession of disconnected single
 workflows). Cylc is unique in its ability to manage these workflows without
 imposing a global cycle loop, so that one cycle does not have to complete
 before the next can start. Instead, Cylc's novel meta-scheduling algorithm runs
-tasks from many cycles at once, to the full extent allowed by individual
-dependencies. So, for example, on restarting after extended downtime, a
-workflow that processes real-time data can clear its backlog and catch up very
-quickly by interleaving cycles.
+cycles concurrently to the full extent allowed by task dependencies. So, for
+example, on restarting after extended downtime, a workflow that processes
+real-time data can clear its backlog and catch up very quickly by interleaving
+cycles.
 
 As a distributed system, Cylc scales sideways: each workflow is managed by its
-own lightweight ad-hoc server program. Existing scripts or programs can be used
+own lightweight server program. Existing scripts or programs can be used
 without modification in Cylc tasks: they are automatically wrapped in code to
 trap errors and report run status via authenticated HTTPS messages or polling.
 Cylc workflows are defined with a graph notation that efficiently expresses
