### 2018-05-29 #737 @hjoliver
@Chilipp @sarats - sorry for the delay, ready for review now as cylc-7.7.0.  I've updated the paper details on the cylc master branch; hopefully that flows through to the JOSS system here...
