### 2018-05-30 3dd4339
Add documentation of new duration filters to CUG.
This builds both the PDF and HTML documentation, but has
some funny effects in the HTML code where the underscores
display properly in the PDF.  Not sure how to address this.
