+++ doc/src/cylc-user-guide/cug.tex
@@ -5161,7 +5161,7 @@ locations:
 
 In the filter function argument list, the first argument is the variable
 value to be ``filtered'', and subsequent arguments can be whatever is
-needed. Currently there are two custom filters:
+needed. Currently there are four custom filters:
 
 
 \paragraph{pad}
@@ -5207,6 +5207,51 @@ Examples:
     \item \lstinline={{'12,30,2000' | strftime('%m', '%m,%d,%Y')}}= - 12
     \item \lstinline={{'1066/10/14 08:00:00' | strftime('%Y%m%dT%H', '%Y/%m/%d %H:%M:%S')}}= - 10661014T08
 \end{myitemize}
+
+\paragraph{duration\_to\_hours}
+
+The ``duration\_to\_hours'' filter can be used to format ISO8601 duration
+strings as a floating-point number of hours.
+
+\lstset{language=suiterc}
+\begin{lstlisting}
+{% set CYCLE_INTERVAL = 'PT1D' %}
+{{ CYCLE_INTERVAL | duration_to_hours }}  # 24.0
+{% set CYCLE_SUBINTERVAL = 'PT30M' %}
+{{ CYCLE_SUBINTERVAL | duration_to_hours }}  # 0.5
+\end{lstlisting}
+
+While the filtered value is a floating-point number, it is often required to
+supply an integer to suite entities (e.g. environment variables) that require
+it.  This is accomplished by chaining filters:
+
+\begin{myitemize}
+    \item \lstinline={{CYCLE_INTERVAL | duration_to_hours | int}}= - 24
+    \item \lstinline={{CYCLE_SUBINTERVAL | duration_to_hours | int}}= - 0
+\end{myitemize}
+
+\paragraph{duration\_to\_seconds}
+
+The ``duration\_to\_seconds'' filter can be used to format ISO8601 duration
+strings as a floating-point number of seconds.
+
+\lstset{language=suiterc}
+\begin{lstlisting}
+{% set CYCLE_INTERVAL = 'PT1D' %}
+{{ CYCLE_INTERVAL | duration_to_seconds }}  # 86400.0
+{% set CYCLE_SUBINTERVAL = 'PT30M' %}
+{{ CYCLE_SUBINTERVAL | duration_to_seconds }}  # 1800.0
+\end{lstlisting}
+
+While the filtered value is a floating-point number, it is often required to
+supply an integer to suite entities (e.g. environment variables) that require
+it.  This is accomplished by chaining filters:
+
+\begin{myitemize}
+    \item \lstinline={{CYCLE_INTERVAL | duration_to_seconds | int}}= - 86400
+    \item \lstinline={{CYCLE_SUBINTERVAL | duration_to_seconds | int}}= - 1800
+\end{myitemize}
+
 
 \subsubsection{Associative Arrays In Jinja2}
 
