### 2018-05-31 d77b53c
modified a python.org and graphviz.org link to be clickable in tex files
