+++ doc/src/cylc-user-guide/cug.tex
@@ -4079,7 +4079,7 @@ object oriented programming languages. The result of this should be
 fairly obvious for typical use of multiple inheritance in cylc suites,
 but for detailed documentation of how the algorithm works refer to the
 official Python documentation here:
-\lstinline=http://www.python.org/download/releases/2.3/mro/=.
+\url{http://www.python.org/download/releases/2.3/mro/}.
 
 The {\em inherit.multi.one} example suite, listed here, makes use of
 multiple inheritance:
