### 2018-06-05 #737 @Chilipp
Hey @hjoliver ! Here are the first points concerning the *General checks*:

- The license file:
   > **License:** Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?

  You state in the [README](https://github.com/cylc/cylc/blob/master/README.md)
  > Cylc is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.

  and
  > You should have received a copy of the GNU General Public License along with cylc.

  but I could not find the LICENSE file in the repository.
- The version:
  > Version: Does the release version given match the GitHub release (cylc-7.6.1)?

  No, it's 7.7.0 now. So I am not sure, how to handle this (pinging @arfon)
