### 2018-06-05 b56f477
Add filter for security warnings from urllib3
Obviously, we are aware of the security issues on the affected
platforms, but these warnings serve no purpose on those platforms except
to confuse or annoy users.
