### 2018-06-05 #737 @Chilipp
Hey @hjoliver , I tried your [Workflow example](https://cylc.github.io/cylc/documentation.html#a-cycling-workflow-example) and it works well (including installation) until I try to run the job. Then it stays in the *submitted* state for the first job.

![image](https://user-images.githubusercontent.com/9960249/40990821-45adf0a4-68f2-11e8-9c3b-8a3e60bb86a0.png)

I have to say, I tested it on my laptop (OSX) without a scheduler. But from what I got in the documentation, that means that I use the *background* scheduler by default which just invokes a bash background job.
So what am I doing wrong here?
