### 2018-06-05 cylc/cylc-flow#2689 @Chilipp
Examples not running under OSX
Refering to https://github.com/openjournals/joss-reviews/issues/737#issuecomment-394786322, I create an issue here because I could not get the software to run on OSX.

After installation and running `cylc check-software`

```
Checking your software...

Individual results:
==========================================================================================
Package (version requirements)                                     Outcome (version found)
==========================================================================================
                                   *REQUIRED SOFTWARE*
Python (2.6+, <3)................................FOUND & min. version MET (2.7.12.final.0)

             *OPTIONAL SOFTWARE for the GUI & dependency graph visualisation*
Python:pygtk (2.0+)......................................FOUND & min. version MET (2.24.0)
graphviz (any)..............................................................FOUND (2.38.0)
Python:pygraphviz (any)......................................................FOUND (1.3.1)

                       *OPTIONAL SOFTWARE for the HTML User Guide*
ImageMagick (any)..........................................................FOUND (6.9.5-0)

                  *OPTIONAL SOFTWARE for the HTTPS communications layer*
Python:urllib3 (any)..........................................................FOUND (1.22)
Python:OpenSSL (any)........................................................FOUND (18.0.0)
Python:requests (2.4.2+).................................FOUND & min. version MET (2.18.4)

                       *OPTIONAL SOFTWARE for the LaTeX User Guide*
TeX:framed (any)...............................................................FOUND (n/a)
TeX (3.0+)...........................................FOUND & min. version MET (3.14159265)
TeX:preprint (any).............................................................FOUND (n/a)
TeX:tex4ht (any)...............................................................FOUND (n/a)
TeX:tocloft (any)..............................................................FOUND (n/a)
TeX:texlive (any)..............................................................FOUND (n/a)
==========================================================================================

Summary:
                               ****************************
                                  Core requirements: ok
                                  Full-functionality: ok
                               ****************************
```
the software is not able to `cylc run examples/7.7.0/tutorial/oneoff/basic`. It stops after having submitted the task

![Screenshot](https://user-images.githubusercontent.com/9960249/40990821-45adf0a4-68f2-11e8-9c3b-8a3e60bb86a0.png)

All the necessary dependencies (besides cylc) have been installed via homebrew.

If I test it on Linux, it runs without problems. Can you reproduce the problem?
