### 2018-06-05 #737 @hjoliver
@Chilipp - yes, cylc can manage background jobs, even on remote hosts.  I'll respond regarding OSX at https://github.com/cylc/cylc/issues/2689
