### 2018-06-05 cylc/cylc#2689 @hjoliver
@Chilipp - thanks for your reviewing efforts so far!

Firstly, good to hear that it runs without problems for you on Linux.

Secondly, the symptom you describe (task job stuck in the "submitted" state) means that the HTTPS job status messages are not getting through to the server program (which is listening on port 43063 according to your screenshot above).

If you look in the `job.err` file for the task you should see some evidence of this.  See `~/cylc-run/my.suite/log/job/2021/prep/01/job.err`

This could mean the port is blocked, or that you don't have ssl libraries installed - unlikely as you're just running on localhost, and your `cylc check-software` results looked good.

More likely (bad news for OSX):  we have been told that it is **possible** to run Cylc on OSX, particularly if your computer has been configured as part of a wider network, **but** something about the default network configuration and/or SSL setup on OSX causes a problem for us.  And this may be difficult to solve, unless you're quite expert on OSX networking.  I believe we have a note about this in the installation section of the user guide.

Unfortunately we on the small development team do not have OSX, we all work on Linux.  So for the moment I'd prefer just to warn users of the potential for problems on OSX.  Hopefully this is not a show-stopper for the review?
