### 2018-06-05 cylc/cylc#2689 @hjoliver
(We should leave this new issue open until the problem is solved, however).
