### 2018-06-06 8c799c3
Restrict urllib3 warnings filter to <2.7.9
Also add extra info to user guide.
