### 2018-06-06 222a638
Fix gscan and profile-battery --help display issue
Invoking the commands with `--help` should no longer require DISPLAY.
