### 2018-06-06 cylc/cylc#2689 @matthewrmshin
(I have modified the user guide as part of #2688.)
