### 2018-06-06 cylc/cylc#2689 @hjoliver
Thanks @matthewrmshin. (I guess we removed the OSX info that used to be in the User Guide; I vaguely recall now, the specific advice was outdated).
