### 2018-06-06 06ff9a5
pbs: handle poll if PBS client cannot connect
If PBS qstat cannot connect to its server, assume that jobs managed by
it are still OK.
