### 2018-06-07 003b9c7
cug: improve statement on python 2 requirement
