+++ doc/src/cylc-user-guide/cug.tex
@@ -387,8 +387,10 @@ including Apple OS X.
 
 \subsection{Third-Party Software Packages}
 
-{\bf Python \lstinline@>=@ 2.6} is required (but not yet Python 3). Python
-should already be installed in your Linux system. \url{https://python.org}.
+{\bf Python 2 \lstinline@>=@ 2.6} is required.
+{\bf Python 2 \lstinline@>=@ 2.7.9} is recommended for the best security.
+Python 2 should already be installed in your Linux system.
+\url{https://python.org/}.
 
 For Cylc's HTTPS communications layer:
 \begin{myitemize}
@@ -396,8 +398,6 @@ For Cylc's HTTPS communications layer:
   \item {\bf pyOpenSSL} - \url{http://www.pyopenssl.org/}
   \item {\bf python-requests} - \url{http://docs.python-requests.org/}
   \item ({\bf python-urllib3} - should be bundled with python-requests)
-  \item ({\bf Python \lstinline@>=@ 2.7.9} (but not yet Python 3) is
-        recommended for the best security.)
 \end{myitemize}
 
 The following packages are highly recommended, but are technically optional as
