### 2018-06-07 cylc/cylc#2694 @sarats
Installation steps issues & suggestions
This is in relation to openjournals/joss-reviews#737.

I have followed the documentation in constructing the simple workflow on a Linux server (CentOS 7.4). I presume it is unable to run due to lack of a background scheduler?

```
$ ./cylc log -o my.suite
2018-06-07T11:26:55-04 INFO - Suite starting: server=n34-003.ftpn.ornl.gov:43076 pid=21409
2018-06-07T11:26:55-04 INFO - Cylc version: 7.7.0
2018-06-07T11:26:55-04 INFO - Run mode: live
2018-06-07T11:26:55-04 INFO - Initial point: 2021
2018-06-07T11:26:55-04 INFO - Final point: 2023
2018-06-07T11:26:55-04 INFO - Cold Start 2021
2018-06-07T11:26:55-04 INFO - [prep.2021] -submit-num=1, owner@host=localhost
2018-06-07T11:26:56-04 INFO - [prep.2021] -(current:ready) submitted at 2018-06-07T11:26:56-04
2018-06-07T11:26:56-04 INFO - [prep.2021] -job[01] submitted to localhost:background[21437]
2018-06-07T11:26:56-04 INFO - [prep.2021] -health check settings: submission timeout=None
$ ./cylc log -o my.suite model.2021
ERROR: file not found: /home/yss/cylc-run/my.suite/log/job/2021/model/NN/job.out
```
The process is still pending in the background. Any suggestions?

I don't have the optional packages on that machine and check-software bails.
```
./bin/cylc check-software
Checking your software...

Individual results:
==========================================================================================
Package (version requirements)                                     Outcome (version found)
==========================================================================================
                                   *REQUIRED SOFTWARE*
Python (2.6+, <3).................................FOUND & min. version MET (2.7.5.final.0)

             *OPTIONAL SOFTWARE for the GUI & dependency graph visualisation*
Python:pygtk (2.0+)......................................FOUND & min. version MET (2.24.0)
graphviz (any)...............................................................NOT FOUND (-)
Traceback (most recent call last):
  File "/home/yss/repos/cylc-7.7.0/bin/cylc-check-software", line 269, in <module>
    main()
  File "/home/yss/repos/cylc-7.7.0/bin/cylc-check-software", line 251, in main
    functionality_print(tag)
  File "/home/yss/repos/cylc-7.7.0/bin/cylc-check-software", line 227, in functionality_print
    opt_result[module] = check_py_module_ver(module, ver_req)
  File "/home/yss/repos/cylc-7.7.0/bin/cylc-check-software", line 154, in check_py_module_ver
    res = ['%s (%s)' % (FOUND_NOVER_MSG, module_version), True]
UnboundLocalError: local variable 'module_version' referenced before assignment

```

Suggestions:

Perhaps the installation procedure and corresponding instructions can be separated for creating version file and generating documentation.  Perhaps a condensed quickstart which just sets version can be considered?

The documentation is readily available on the website and one just needs to update the version info in their local directory.  The documentation step requires various LaTeX extensions etc. which need not/cannot be installed easily on a HPC platform.  Anyway, these missing dependencies can mislead someone that the installation isn't ready to use when one types 'make'. Perhaps remove documentation from the 'all' target in Makefile?

I have also tried this on a Linux laptop running RHEL 7.5 where I could install some of the optional dependencies.

Finally, I would like to commend the nice effort behind the comprehensive documentation.
