### 2018-06-07 cylc/cylc#2694 @matthewrmshin
Note: The `cylc check-software` traceback should be fixed by #2683.
