### 2018-06-07 cylc/cylc#2694 @matthewrmshin
Perhaps it is a bit unclear in our docs, but a background submission means that the job is run as a Unix process in the background, so it has probably submitted fine but is stuck trying to communicate back to the suite server program. An issue with the network set up is most likely.

Do you mind printing out the STDERR of the job to give us more info?

```bash
./cylc log -e my.suite prep.2021
# OR
cat ~/cylc-run/my.suite/log/job/2021/prep/NN/job.err
```

(We run Cylc on RHEL 6 and 7 systems on our site, so I am confident that we can make it work for you as well.)
