### 2018-06-07 cylc/cylc#2694 @sarats
@matthewrmshin Yes, it is a problem with the restrictive network.
```
$ cat ~/cylc-run/my.suite/log/job/2021/prep/NN/job.err
2018-06-07T11:26:58-04 WARNING - Message send failed, try 1 of 7: Cannot connect: https://<myhost>:43076/put_messages: HTTPSConnectionPool(host='myhost.ornl.gov', port=43076): Max retries exceeded with url: /put_messages (Caused by ProxyError('Cannot connect to proxy.', error('Tunnel connection failed: 403 Forbidden',)))
```

I was able to run successfully on my Linux laptop.
```
$ ./bin/cylc log -o my.suite model.2021
Suite    : my.suite
Task Job : 2021/model/01 (try 1)
User@Host: sarat@<myhost>

my FOOD is icecream
2018-06-07T16:35:24-04 INFO - started
```
