### 2018-06-07 cylc/cylc#2694 @sarats
@matthewrmshin Good to know about check-software fix.
