### 2018-06-07 cylc/cylc#2694 @matthewrmshin
Thanks for investigating. I wonder if you can try disabling the proxy when running Cylc. E.g. Is the proxy setting defined by an environment variable such as `https_proxy`? If so, I wonder if it will work if you unset the environment variable or export `no_proxy` to point to the current host?
