### 2018-06-08 cylc/cylc#2689 @Chilipp
Hi @hjoliver and @matthewrmshin! Thanks for your quick response to that.

I think that is a good solution, I would just phrase it a bit differently to be more transparent (see [the code comment](https://github.com/cylc/cylc/pull/2688/files/8c799c33353a02617dfc8f044b61702ba41b15f5#r194048585) in PR https://github.com/cylc/cylc/pull/2688)
