### 2018-06-08 5191663
cug: improve statement on Linux requirement
Make it explicit that Apple OSX is not officially tested.
