+++ doc/src/cylc-user-guide/cug.tex
@@ -383,7 +383,7 @@ previous section, emerges naturally at run time.
 
 Cylc runs on Linux. It is tested quite thoroughly on modern RHEL and Ubuntu
 distros. Some users have also managed to make it work on other Unix variants
-including Apple OS X.
+including Apple OS X, but they are not officially tested and supported.
 
 \subsection{Third-Party Software Packages}
 
