### 2018-06-08 67f6645
Fix all genuine amendable Python issues picked up by Codacy
