### 2018-06-08 #737 @Chilipp
Dear all,

I am done now with my review so far, went through the user guide and thanks a lot for this nice and well-developed open source package. I find it well documented and helpful and therefore, if the points already mentioned previously (version https://github.com/openjournals/joss-reviews/issues/737#issuecomment-394667300 and contribution guide https://github.com/openjournals/joss-reviews/issues/737#issuecomment-395481912) are fixed and commented by @arfon, I recommend to accept the paper.

Personally, however, there are two points I would like to mention:

1. I am not happy about the reliance on Python2 and GTK, but as you already state in the paper, this is already on the roadmap and this is anyway no reason for JOSS to reject the paper.
2. Concerning the tests on [Travis CI](https://travis-ci.org/cylc/cylc) I wonder, how you are tracking how well your software is covered by the tests? The core of it is a python module, so it should be possible to some extent at least using [coverage](https://coverage.readthedocs.io/en/coverage-4.5.1/) or some equivalent tool.
   It would be very helpful (for reviewing and for you as a developer) to have a coverage report. But of course, due to the complexity and scale of the software, I understand that this is not completely straight-forward and anyway, the questions above just ask
   > Are there automated tests or manual steps described so that the function of the software can be verified?

   So this is definitely fulfilled and, additionally, it seems to me that your test suite covers all the central parts of the software, tutorials, etc.. Therefore I recommend to ignore this shortcoming.
