### 2018-06-09 #737 @hjoliver
@sarats - thank you. In reponse to your points:
 * LICENSE vs COPYING - let's see the response from @arfon
 * ditto - need advice from @arfon on how to correct the version in the submission
 * lack of CONTRIBUTING.md is a problem - we have planned to correct this since some months ago, but still waiting on legal advice on that from my home institution (I will chase them up...)
 * we have Tracis CI hooked up to the main cylc repository on GitHub - every new commit triggers an increasingly massive battery of tests (currently takes almost 50 minutes to run, typically). Similar Codacy for static source code analysis. Code coverage testing is yet to be investigated and tested, but it is on my radar.
