### 2018-06-11 #737 @arfon
> - LICENSE info is in COPYING file as per GPL v3 whereas the Journal requires a LICENSE file? @arfon Can you clarify?

`COPYING` is acceptable, especially as this is recommended by the GPL.

> - Version needs to be fixed in submission as pointed by @Chilipp.

I can update this. What should the new version be?
