### 2018-06-11 #737 @hjoliver
Thanks @arfon

New version (as reviewed) is 7.7.0
