### 2018-06-11 #737 @sarats
> * lack of CONTRIBUTING.md is a problem - we have planned to correct this since some months ago, but still waiting on legal advice on that from my home institution (I will chase them up...)

@hjoliver I understand how complex formulating this policy can be. I'm involved a U.S. Dept. of Energy funded multi-lab project where we encourage people willing to contribute to submit bug reports or fixes for simple scenarios.  https://github.com/E3SM-Project/E3SM/blob/master/CONTRIBUTING.md
However we recommend a formal collaboration plan for more extensive changes.

Food for thought: You can perhaps have something simple for the purpose of this review before finalizing your policy for extensive IP transfer etc.?
Anyway, this is the only blocker.

Otherwise this submission is good to go. Thanks for all the hard work.
