### 2018-06-11 #737 @hjoliver
Thanks very much @sarats and @Chilipp (and @arfon).  I'll discuss with the team in the next few days what we can do about a preliminary contributing.md ...
