### 2018-06-22 0d25621
Assorted degbugging, including of the utility tests themselves
