+++ lib/cylc/rundb.py
@@ -987,6 +987,55 @@ class CylcReviewDAO(object):
         "succeeded,failed": "run_status IS NOT NULL",
         "failed": "run_status == 1",
     }
+    JOB_ORDERS = {
+        "time_desc": "time DESC, submit_num DESC, name DESC, cycle DESC",
+        "time_asc": "time ASC, submit_num ASC, name ASC, cycle ASC",
+        "cycle_desc_name_asc": "cycle DESC, name ASC, submit_num DESC",
+        "cycle_desc_name_desc": "cycle DESC, name DESC, submit_num DESC",
+        "cycle_asc_name_asc": "cycle ASC, name ASC, submit_num DESC",
+        "cycle_asc_name_desc": "cycle ASC, name DESC, submit_num DESC",
+        "name_asc_cycle_asc": "name ASC, cycle ASC, submit_num DESC",
+        "name_desc_cycle_asc": "name DESC, cycle ASC, submit_num DESC",
+        "name_asc_cycle_desc": "name ASC, cycle DESC, submit_num DESC",
+        "name_desc_cycle_desc": "name DESC, cycle DESC, submit_num DESC",
+        "time_submit_desc": (
+            "time_submit DESC, submit_num DESC, name DESC, cycle DESC"),
+        "time_submit_asc": (
+            "time_submit ASC, submit_num DESC, name DESC, cycle DESC"),
+        "time_run_desc": (
+            "time_run DESC, submit_num DESC, name DESC, cycle DESC"),
+        "time_run_asc": (
+            "time_run ASC, submit_num DESC, name DESC, cycle DESC"),
+        "time_run_exit_desc": (
+            "time_run_exit DESC, submit_num DESC, name DESC, cycle DESC"),
+        "time_run_exit_asc": (
+            "time_run_exit ASC, submit_num DESC, name DESC, cycle DESC"),
+        "duration_queue_desc": (
+            "(CAST(strftime('%s', time_run) AS NUMERIC) -" +
+            " CAST(strftime('%s', time_submit) AS NUMERIC)) DESC, " +
+            "submit_num DESC, name DESC, cycle DESC"),
+        "duration_queue_asc": (
+            "(CAST(strftime('%s', time_run) AS NUMERIC) -" +
+            " CAST(strftime('%s', time_submit) AS NUMERIC)) ASC, " +
+            "submit_num DESC, name DESC, cycle DESC"),
+        "duration_run_desc": (
+            "(CAST(strftime('%s', time_run_exit) AS NUMERIC) -" +
+            " CAST(strftime('%s', time_run) AS NUMERIC)) DESC, " +
+            "submit_num DESC, name DESC, cycle DESC"),
+        "duration_run_asc": (
+            "(CAST(strftime('%s', time_run_exit) AS NUMERIC) -" +
+            " CAST(strftime('%s', time_run) AS NUMERIC)) ASC, " +
+            "submit_num DESC, name DESC, cycle DESC"),
+        "duration_queue_run_desc": (
+            "(CAST(strftime('%s', time_run_exit) AS NUMERIC) -" +
+            " CAST(strftime('%s', time_submit) AS NUMERIC)) DESC, " +
+            "submit_num DESC, name DESC, cycle DESC"),
+        "duration_queue_run_asc": (
+            "(CAST(strftime('%s', time_run_exit) AS NUMERIC) -" +
+            " CAST(strftime('%s', time_submit) AS NUMERIC)) ASC, " +
+            "submit_num DESC, name DESC, cycle DESC"),
+        }
+
     REC_CYCLE_QUERY_OP = re.compile(r"\A(before |after |[<>]=?)(.+)\Z")
[…]
