### 2018-06-24 7c879b7
Address review feedback on xtriggers.
- don't store clock offset in taskdef
- point not needed in wall_clock
- default zero offset in wall_clock
- predefined zero offset wall_clock
- allow positional and non-string args
