### 2018-06-24 #737 @hjoliver
(I've been given the legal go-ahead for a contributor agreement, will try to add something up this week).
