### 2018-06-24 6985909
Use json not pickle for xtrig db table.
