### 2018-06-25 1d0bb4c
Pep8; restore travis config.
