+++ .travis.yml
@@ -58,7 +58,6 @@ install:
     # Pygraphviz needs special treatment to avoid an error from "from . import release"
     - pip install pygraphviz --install-option="--include-path=/usr/include/graphviz" --install-option="--library-path=/usr/lib/graphviz/"
     - pip install pycodestyle pyopenssl
-    - pip install kafka  # just for validating dev/suites/xtrigger/kafka/
 
 # Run tests
 script: 
@@ -67,10 +66,9 @@ script:
     # Only run the generic tests on Travis CI.
     - export CYLC_TEST_RUN_PLATFORM=false
     # Run tests with virtual frame buffer for X support.
-    #- xvfb-run -a cylc test-battery --state=save -j 5 || (echo -e "\n\nRerunning Failed Tests...\n\n"; cylc test-battery --state=save,failed -j 5)
-    - xvfb-run -a cylc test-battery tests/syntax tests/xtriggers
+    - xvfb-run -a cylc test-battery --state=save -j 5 || (echo -e "\n\nRerunning Failed Tests...\n\n"; cylc test-battery --state=save,failed -j 5)
 
 # Check output (more useful if you narrow down what tests get run)
-#after_script:
-#    - find $HOME/cylc-run -name '*.err' -type f -exec echo '==== {} ====' \; -exec cat '{}' \;
-#    - find /tmp/${USER}/cylctb-* -type f -exec echo '==== {} ====' \; -exec cat '{}' \;
+after_script:
+    - find $HOME/cylc-run -name '*.err' -type f -exec echo '==== {} ====' \; -exec cat '{}' \;
+    - find /tmp/${USER}/cylctb-* -type f -exec echo '==== {} ====' \; -exec cat '{}' \;
