### 2018-06-26 4f5ed67
Better predefined wall_clock trigger impl.
