### 2018-06-26 ae4f381
integer column does not accept None as a value
