### 2018-06-27 d22b99a
Updated change log for 7.7.1
