+++ CHANGES.md
@@ -6,6 +6,50 @@ milestones](https://github.com/cylc/cylc/milestones?state=closed) for each
 release.
 
 -------------------------------------------------------------------------------
+## __cylc-7.7.1 (2018-06-27)__
+
+Minor maintenance release.
+
+### Fixes
+
+(Several minor fixes have been omitted from this list.)
+
+[#2678](https://github.com/cylc/cylc/pull/2678) - fix loading of job poll
+timers on restart (bug introduced at last release)
+
+[#2683](https://github.com/cylc/cylc/pull/2683) - fix potential error in
+`cylc check-software` (which checks for installed software dependencies)
+
+[#2691](https://github.com/cylc/cylc/pull/2691) PBS support - handle job poll
+result correctly if qstat temporarily fails to connect to the server
+
+[#2703](https://github.com/cylc/cylc/pull/2703) - fix an error
+(inconsequential) that appears in the suite log at restart: `ValueError: No
+JSON object could be decoded`
+
+[#2692](https://github.com/cylc/cylc/pull/2692) - fix X11 forwarding for GUI
+edit job log, with `cylc gui --host=HOST`
+
+[#2690](https://github.com/cylc/cylc/pull/2690) - invoking Cylc command help
+should not require `$DISPLAY` to be set
+
+[#2677](https://github.com/cylc/cylc/pull/2677) - use random serial numbers in
+the self-signed SSL certificates generated by suite server programs
+
+[#2688](https://github.com/cylc/cylc/pull/2688) 
+[#2705](https://github.com/cylc/cylc/pull/2705) - block several security
+warnings emitted by `urllib3` under old Python versions (2.6). *We are
+aware of the security issues, but these warnings serve no purpose on affected
+platforms except to confuse and annoy users.*
+
+[#2676](https://github.com/cylc/cylc/pull/2676) - use `#!/usr/bin/env python2`
+(i.e. Python-2 specific) in Cylc source files, to avoid issues with default
+Python 3 installations (note Cylc is going to Python 3 next year)
+
+[#2679](https://github.com/cylc/cylc/pull/2679) - change bold font back to
+normal in the GUI log viewer
+
+-------------------------------------------------------------------------------
 ## __cylc-7.7.0 (2018-05-12)__
 
 ### Enhancements
