+++ documentation.md
@@ -21,6 +21,10 @@ See [FAQ](./faq.html).
 If you have access to cylc already, type `cylc doc` or use the GUI "Help" menu
 to view the User Guide.  Otherwise, an online copy is available here:
 
+_Warning: the auto-generated HTML version of the user guide is quite ugly and
+has some formatting issues; we plan to reformulate the Cylc documentation using
+Sphinx in due course, which will fix this problem._
+
 * [Cylc User Guide - PDF ~3.5MB](doc/cylc-user-guide.pdf)
 * [Cylc User Guide - HTML single page](html/single/cug-html.html)
 * [Cylc User Guide - HTML multi page](html/multi/cug-html.html)
