### 2018-06-27 7fc05e4
Branch code tidy.
