### 2018-06-28 f0f438b
CUG - document xtriggers.
