+++ doc/src/cylc-user-guide/suiterc.tex
@@ -3,7 +3,7 @@
 
 \lstset{language=bash}
 
-This appendix defines all legal suite definition config items.
+This appendix defines all legal suite configuration items.
 Embedded Jinja2 code (see~\ref{Jinja2}) must process to a valid
 raw suite.rc file. See also~\ref{SuiteRCFile} for a descriptive
 overview of suite.rc files, including syntax (\ref{Syntax}).
@@ -950,10 +950,14 @@ see~\ref{ClockExpireTasks}.
 
 \paragraph[external-trigger]{[scheduling] \textrightarrow [[special tasks]] \textrightarrow external-trigger}
 
-Externally triggered tasks (see~\ref{ExternalTriggers}) wait on external events
-reported via the \lstinline=cylc ext-trigger= command.  To constrain triggers
-to a specific cycle point, include \lstinline=$CYLC_TASK_CYCLE_POINT= in the
-trigger message string and pass the cycle point to the \lstinline=cylc ext-trigger= command.
+{\bf This feature is now deprecated. Please use {\em external trigger
+functions} instead - see~\ref{External Trigger Functions}}
+
+Externally triggered tasks (see~\ref{Old-Style External Triggers}) wait on
+external events reported via the \lstinline=cylc ext-trigger= command. To
+constrain triggers to a specific cycle point, include
+\lstinline=$CYLC_TASK_CYCLE_POINT= in the trigger message string and pass the
+cycle point to the \lstinline=cylc ext-trigger= command.
 
 \begin{myitemize}
     \item {\em type:} Comma-separated list of task names with associated
@@ -1595,17 +1599,17 @@ delay.
 
 \subparagraph[suite definition directory]{[runtime] \textrightarrow [[\_\_NAME\_\_]] \textrightarrow [[[remote]]] \textrightarrow  suite definition directory}
 
-The path to the suite definition directory on the remote account, needed if
+The path to the suite configuration directory on the remote account, needed if
 remote tasks require access to files stored there (via
 \lstinline=$CYLC_SUITE_DEF_PATH=) or in the suite bin directory (via
 \lstinline=$PATH=).  If this item is not defined, the local suite
-definition directory path will be assumed, with the suite owner's home
+configuration directory path will be assumed, with the suite owner's home
 directory, if present, replaced by \lstinline='$HOME'= for
 interpretation on the remote account.
 
 \begin{myitemize}
 \item {\em type:} string (a valid directory path on the remote account)
-\item {\em default:} (local suite definition path with \lstinline=$HOME=
+\item {\em default:} (local suite configuration path with \lstinline=$HOME=
     replaced)
 \end{myitemize}
 
@@ -2270,7 +2274,7 @@ Here you can assign graph node attributes to specific nodes, or to all
 members of named groups defined in [visualization] \textrightarrow [[node
 groups]]. Task families are automatically node groups. Styling of a
 family node applies to all member nodes (tasks and sub-families), but
-precedence is determined by ordering in the suite definition.  For
+precedence is determined by ordering in the suite configuration.  For
 example, if you style a family red and then one of its members green,
 cylc will plot a red family with one green member; but if you style one
 member green and then the family red, the red family styling will
