+++ doc/src/cylc-user-guide/cug.tex
@@ -689,7 +689,7 @@ documented in (\ref{SiteRCReference}).
 # cylc site global config file
 <cylc-dir>/etc/global.rc
 \end{lstlisting}
-Others, such as the preferred text editor for suite definitions,
+Others, such as the preferred text editor for suite configurations,
 can be overridden by users,
 \lstset{language=transcript}
 \begin{lstlisting}
@@ -712,7 +712,7 @@ functionality.
 
 Some settings affecting cylc's behaviour can be defined in site and user
 {\em global config files}. For example, to choose the text editor invoked by
-cylc on suite definitions:
+cylc on suite configurations:
 
 \lstset{language=suiterc}
 \begin{lstlisting}
@@ -774,15 +774,15 @@ $ cylc gscan & # Multi-suite monitor GUI.
 Clicking on a suite in gscan, shown in Figure~\ref{fig-gscan}, opens a gcylc
 instance for it.
 
-\subsection{Suite Definitions}
+\subsection{Suite Configuration}
 
 Cylc suites are defined by extended-INI format \lstinline=suite.rc=
 files (the main file format extension is section nesting). These reside
-in {\em suite definition directories} that may also contain a
+in {\em suite configuration directories} that may also contain a
 \lstinline=bin= directory and any other suite-related files.
 
 \begin{myitemize}
-\item For more on the suite definition file format, see~\ref{SuiteDefinition}
+\item For more on the suite configuration file format, see~\ref{SuiteDefinition}
     and~\ref{SuiteRCReference}.
 \end{myitemize}
 
@@ -790,7 +790,7 @@ in {\em suite definition directories} that may also contain a
 
 Suite registration creates a run directory (under \lstinline=~/cylc-run/= by
 default) and populates it with authentication files and a symbolic link to a
-suite definition directory. Cylc commands that parse suite definitions can take
+suite configuration directory. Cylc commands that parse suites can take
 the file path or the suite name as input. Commands that interact with running
 suites have to target the suite by name.
 
@@ -857,7 +857,7 @@ See \lstinline=cylc print --help= for other display options.
 
 \subsection{Suite Validation}
 
-Suite definitions can be validated to detect syntax (and other) errors:
+Suite configurations can be validated to detect syntax (and other) errors:
 
 \lstset{language=transcript}
 \begin{lstlisting}
@@ -909,11 +909,10 @@ information and exits).
 
 \subsection{Editing Suites}
 
-The text editor invoked by cylc on suite definitions is determined
+The text editor invoked by Cylc on suite configurations is determined
 by cylc site and user global config files, as shown above in~\ref{CUI}.
 Check that you have renamed the tutorial examples suites as described
-just above and open the {\em Hello World} suite definition in your text
-editor:
+just above and open the {\em Hello World} suite in your text editor:
 \lstset{language=transcript}
 \begin{lstlisting}
 $ cylc edit tut/oneoff/basic # in-terminal
@@ -926,7 +925,7 @@ $ gcylc tut/oneoff/basic &
 \end{lstlisting}
 and choose {\em Suite } \textrightarrow {\em Edit} from the menu.
 
-The editor will be invoked from within the suite definition directory for easy
+The editor will be invoked from within the suite configuration directory for easy
 access to other suite files (in this case there are none). There are syntax
 highlighting control files for several text editors under
 \lstinline=<cylc-dir>/etc/syntax/=; see in-file comments for installation
@@ -999,7 +998,7 @@ $ cylc gui tut/oneoff/basic &
 \end{lstlisting}
 Use the tool bar {\em Play} button, or the {\em Control}
 \textrightarrow {\em Run} menu item, to run the suite again.
-You may want to alter the suite definition slightly to make the task
+You may want to alter the suite configuration slightly to make the task
[…]
