### 2018-06-28 a4e0fa4
Various minor tweaks (feedback and codacy).
