+++ doc/src/cylc-user-guide/cug.tex
@@ -6328,6 +6328,8 @@ If necessary the filesystem can be used for this purpose.}
   \item recommend using built-in trigger functions as an example 
   \item refer above for result qualification by label name 
   \item show another example (the random and echo built-ins?)
+  \item example graphs
+  \item debugging trigger functions
 \end{myitemize}
 
 \subsection{Current Limitations}
