+++ doc/src/cylc-user-guide/cug.tex
@@ -3688,12 +3688,10 @@ other tasks beyond the final point.
 \paragraph{Clock Triggers}
 \label{ClockTriggerTasks}
 
-{\em NOTE: this is the original Cylc clock-triggering mechanism in which the
-time offset relative to cycle point is stored as an attribute of the tasks that
-depend on it. It is deprecated in favour of the newer built in
-\lstinline=wall_clock= external trigger function, which is more efficient
-(trigger functions can be shared by many tasks) and is conveniently exposed in
-the graph).}
+{\em NOTE: this older clock-triggering mechanism is now deprecated in favour of
+built-in \lstinline=wall_clock= external triggers, which are more transparent
+(visible in the graph) and efficient (triggeris are shared among dependent
+tasks).}
 
 By default, date-time cycle points are not connected to the real time ``wall
 clock''. They are just labels that are passed to task jobs (e.g.\ to
@@ -6097,87 +6095,64 @@ hosts} in one of these locations:
 \section{External Event Triggers}
 \label{External Event Triggers}
 
-External triggers allow tasks to trigger directly off of external events, which
-is often preferable to implementing long-running tasks in the workflow that
-poll for external events. The triggering mechanism described in this section
-replaces an older and less powerful one documented in~\ref{Old-Style
-External Triggers}.
-
-{\em WARNING - the suite configuration interface for this new capability
-may change somewhat in future releases - see {\em Current Limitations} below
+{\em WARNING: this is a new capability - its suite configuration interface may
+change somewhat in future releases - see {\em Current Limitations} below
 in~\ref{Current Trigger Function Limitations}.}
+
+External triggers allow tasks to trigger directly off of external events, which
+is often preferable to implementing long-running polling tasks in the workflow.
+The triggering mechanism described in this section replaces an older and less
+powerful one documented in~\ref{Old-Style External Triggers}.
 
 If you can write a Python function to check the status of an external
 condition or event, the suite server program can call it at configurable
 intervals until it reports success, at which point dependent tasks can trigger
 and data returned by the function will be passed to the job environments of
 those tasks. Functions can be written for triggering off of almost anything,
-such as delivery of a new dataset, creation of new entries in some database
+such as delivery of a new dataset, creation of a new entry in a database
 table, or appearance of new data availability notifications in a message
 broker.
 
-Cylc has built-in trigger functions for suite state polling (i.e.\ for
-inter-suite triggering, see~\ref{Built-in Suite State Triggers}) and for clock
-triggering (see~\ref{Built-in Clock Triggers}).
+External triggers are visible in suite visualizations as bare graph nodes (just
+the trigger names). They are plotted against all dependent tasks, not in a
+cycle point specific way like tasks. This is because external triggers may or 
+may not be cycle point (or even task name) specific - it depends on the
+arguments passed to the corresponding trigger functions. For example, if an
+external trigger does not depend on task name or cycle point it will only be
+called once - albeit repeatedly until satisfied - for the entire suite run,
+after which the function result will be remembered for all dependent tasks
+throughout the suite run.
 
-Trigger functions are just normal Python functions with the following
-properties and constraints:
+Several built-in external trigger functions are located in
+\lstinline=<cylc-dir>/lib/cylc/xtriggers/=:
 \begin{myitemize}
-  \item they must be defined in a module with the same name as the function
-  \item they can be located in:
-    \begin{myitemize}
-      \item \lstinline=<cylc-dir>/lib/cylc/xtriggers/=
[…]
