### 2018-06-29 e6c1fbd
Fix doc typos. [skip ci]
