+++ doc/src/cylc-user-guide/suiterc.tex
@@ -918,8 +918,7 @@ declarations and corresponding labels for use in the graph:
 
 \begin{myitemize}
 \item {\em type:} string: function signature followed by optional call interval
-\item {\em example:} \lstinline@trig_1 = my_trigger(arg1, arg2, kwarg1,
-  kwarg2):PT10S@
+\item {\em example:} \lstinline@trig_1 = my_trigger(arg1, arg2, kwarg1, kwarg2):PT10S@
 \end{myitemize}
 
 (See~\ref{External Event Triggers} for details).
