+++ doc/src/cylc-user-guide/cug.tex
@@ -3111,7 +3111,7 @@ be made to trigger off of the state of other tasks (indicated by a
 name in the graph) and, and off the clock, and arbitrary external events.
 
 External triggering is relatively more complicated, and is documented
-separately in Section~\ref{External Triggering}.
+separately in Section~\ref{External Event Triggers}.
 
 \paragraph{Success Triggers}
 
