### 2018-06-29 e00eedc
Final branch docs and hot fixes.
- xclock doesn't need get_func (wall_clock function is built-in)
- fix xtrigger module import via get_func, for local lib/python
- xtrig debug mode via string template
- fix unnecessary lambda in kafka example.
- test: load local xtrigger automatically.
- final branch doc tweaks.
