+++ lib/cylc/scheduler.py
@@ -208,7 +208,7 @@ class Scheduler(object):
         self.broadcast_mgr = BroadcastMgr(self.suite_db_mgr)
         self.xtrigger_mgr = XtriggerManager(
             self.suite, self.owner, self.broadcast_mgr, self.suite_run_dir,
-            self.suite_share_dir, self.suite_work_dir)
+            self.suite_share_dir, self.suite_work_dir, self.suite_dir)
         self.suite_log = None
         self.ref_test_allowed_failures = []
         # Last 10 durations (in seconds) of the main loop
