+++ doc/src/cylc-user-guide/suiterc.tex
@@ -908,8 +908,8 @@ A list of member tasks, or task family names, to assign to this queue
 
 \subsubsection[{[[}xtriggers{]]}]{[scheduling] \textrightarrow [[xtriggers]]}
 
-This section is for {\em External Event Trigger} function declarations -
-see~\ref{External Event Triggers}.
+This section is for {\em External Trigger} function declarations -
+see~\ref{External Triggers}.
 
 \paragraph[\_\_MANY\_\_]{[scheduling] \textrightarrow [[xtriggers]] \textrightarrow \_\_MANY\_\_}
 
@@ -921,7 +921,7 @@ declarations and corresponding labels for use in the graph:
 \item {\em example:} \lstinline@trig_1 = my_trigger(arg1, arg2, kwarg1, kwarg2):PT10S@
 \end{myitemize}
 
-(See~\ref{External Event Triggers} for details).
+(See~\ref{External Triggers} for details).
 
 \subsubsection[{[[}special tasks{]]}]{[scheduling] \textrightarrow [[special tasks]]}
 
@@ -930,8 +930,8 @@ be used in special task lists as shorthand for listing all member tasks.
 
 \paragraph[clock-trigger]{[scheduling] \textrightarrow [[special tasks]] \textrightarrow clock-trigger}
 
-{\em This feature is deprecated, please use {\em Built-in Clock Triggers}
-instead - see~\ref{Built-in Clock Triggers}.}
+{\em NOTE: please read {\em External Triggers} (\ref{External Triggers}) before
+using the older clock triggers described in this section.} 
 
 Clock-trigger tasks (see~\ref{ClockTriggerTasks}) wait on a wall clock time
 specified as an offset from their own cycle point.
@@ -970,8 +970,8 @@ see~\ref{ClockExpireTasks}.
 
 \paragraph[external-trigger]{[scheduling] \textrightarrow [[special tasks]] \textrightarrow external-trigger}
 
-{\em This feature is deprecated, please use {\em External Event Triggers}
-instead - see~\ref{External Event Triggers}.}
+{\em NOTE: please read {\em External Triggers} (\ref{External Triggers}) before
+using the older mechanism described in this section.} 
 
 Externally triggered tasks (see~\ref{Old-Style External Triggers}) wait on
 external events reported via the \lstinline=cylc ext-trigger= command. To

[…]
