### 2018-06-30 c857ea3
Fix a few typos
