### 2018-07-01 b35f458
Final_2 branch doc tweaks. [skip ci]
