### 2018-07-02 03bc445
Port security PR and fix text error inherited from Bush
