### 2018-07-02 355f92a
Tidy-up & final pre-review debug
