+++ lib/cylc/rundb.py
@@ -28,7 +28,7 @@ from glob import glob
 
 import cylc.flags
 from cylc.wallclock import get_current_time_string
-from cylc.suite_logging import LOG, ERR
+from cylc.suite_logging import LOG
 from cylc.task_state import TASK_STATUS_GROUPS
 
 
@@ -503,8 +503,8 @@ class CylcSuiteDAO(object):
         select from broadcast_states table if id_key == CHECKPOINT_LATEST_ID.
         Otherwise select from broadcast_states_checkpoints where id == id_key.
         """
-        stmt, stmt_args = pre_select_broadcast_states(self, id_key=None,
-                                                      order=sort)
+        stmt, stmt_args = self.pre_select_broadcast_states(id_key=None,
+                                                           order=sort)
         for row_idx, row in enumerate(self.connect().execute(stmt, stmt_args)):
             callback(row_idx, list(row))
 
@@ -523,7 +523,7 @@ class CylcSuiteDAO(object):
         Invoke callback(row_idx, row) on each row, where each row contains:
             [time, change, point, namespace, key, value]
         """
-        stmt, stmt_args = pre_select_broadcast_events(self, order=sort)
+        stmt, stmt_args = self.pre_select_broadcast_events(order=sort)
         for row_idx, row in enumerate(self.connect().execute(stmt, stmt_args)):
             callback(row_idx, list(row))
 
@@ -1044,18 +1044,7 @@ class CylcSuiteDAO(object):
 class CylcReviewDAO(object):
     """Cylc Review data access object to the suite runtime database."""
 
-    JOB_STATUS_COMBOS = {
-        "all": "",
-        "submitted": "submit_status == 0 AND time_run IS NULL",
-        "submitted,running": "submit_status == 0 AND run_status IS NULL",
-        "submission-failed": "submit_status == 1",
-        "submission-failed,failed": "submit_status == 1 OR run_status == 1",
-        "running": "time_run IS NOT NULL AND run_status IS NULL",
-        "running,succeeded,failed": "time_run IS NOT NULL",
-        "succeeded": "run_status == 0",
-        "succeeded,failed": "run_status IS NOT NULL",
-        "failed": "run_status == 1",
-    }
+    CYCLE_ORDERS = {"time_desc": " DESC", "time_asc": " ASC"}
     JOB_ORDERS = {
         "time_desc": "time DESC, submit_num DESC, name DESC, cycle DESC",
         "time_asc": "time ASC, submit_num ASC, name ASC, cycle ASC",
@@ -1103,7 +1092,19 @@ class CylcReviewDAO(object):
             "(CAST(strftime('%s', time_run_exit) AS NUMERIC) -" +
             " CAST(strftime('%s', time_submit) AS NUMERIC)) ASC, " +
             "submit_num DESC, name DESC, cycle DESC"),
-        }
+    }
+    JOB_STATUS_COMBOS = {
+        "all": "",
+        "submitted": "submit_status == 0 AND time_run IS NULL",
+        "submitted,running": "submit_status == 0 AND run_status IS NULL",
+        "submission-failed": "submit_status == 1",
+        "submission-failed,failed": "submit_status == 1 OR run_status == 1",
+        "running": "time_run IS NOT NULL AND run_status IS NULL",
+        "running,succeeded,failed": "time_run IS NOT NULL",
+        "succeeded": "run_status == 0",
+        "succeeded,failed": "run_status IS NOT NULL",
+        "failed": "run_status == 1",
+    }
 
     REC_CYCLE_QUERY_OP = re.compile(r"\A(before |after |[<>]=?)(.+)\Z")
[…]
