### 2018-07-03 10061ee
Misc relocation and fixes
Relocate dependency on wallclock module.
Rename subprocess context classes.
