+++ lib/cylc/state_summary_mgr.py
@@ -18,10 +18,10 @@
 """Manage suite state summary for client, e.g. GUI."""
 
 from time import time
+from wallclock import (
+    TIME_ZONE_LOCAL_INFO, TIME_ZONE_UTC_INFO, get_utc_mode)
 
-import cylc.flags
 from cylc.task_id import TaskID
-from cylc.wallclock import TIME_ZONE_LOCAL_INFO, TIME_ZONE_UTC_INFO
 from cylc.suite_status import (
     SUITE_STATUS_HELD, SUITE_STATUS_STOPPING,
     SUITE_STATUS_RUNNING, SUITE_STATUS_RUNNING_TO_STOP,
@@ -118,7 +118,7 @@ class StateSummaryMgr(object):
                 global_summary[key] = str(value)
             else:
                 global_summary[key] = None
-        if cylc.flags.utc:
+        if get_utc_mode():
             global_summary['time zone info'] = TIME_ZONE_UTC_INFO
         else:
             global_summary['time zone info'] = TIME_ZONE_LOCAL_INFO
