+++ lib/cylc/network/httpserver.py
@@ -31,6 +31,8 @@ import traceback
 from uuid import uuid4
 
 import cherrypy
+from wallclock import RE_DATE_TIME_FORMAT_EXTENDED
+
 from cylc.cfgspec.glbl_cfg import glbl_cfg
 from cylc.exceptions import CylcError
 import cylc.flags
@@ -43,7 +45,6 @@ from cylc.suite_srv_files_mgr import (
     SuiteSrvFilesManager, SuiteServiceFileError)
 from cylc.unicode_util import utf8_enforce
 from cylc.version import CYLC_VERSION
-from cylc.wallclock import RE_DATE_TIME_FORMAT_EXTENDED
 
 
 class HTTPServer(object):
