+++ lib/cylc/task_state.py
@@ -18,6 +18,8 @@
 
 """Task state related logic."""
 
+from wallclock import get_current_time_string
+
 import cylc.flags as flags
 from cylc.prerequisite import Prerequisite
 from cylc.suite_logging import LOG
@@ -26,7 +28,6 @@ from cylc.task_outputs import (
     TaskOutputs,
     TASK_OUTPUT_EXPIRED, TASK_OUTPUT_SUBMITTED, TASK_OUTPUT_SUBMIT_FAILED,
     TASK_OUTPUT_STARTED, TASK_OUTPUT_SUCCEEDED, TASK_OUTPUT_FAILED)
-from cylc.wallclock import get_current_time_string
 
 
 # Task status names and meanings.

[…]
