### 2018-07-03 a7b9d9a
Fix urlib3 warning with python >= 2.7.9
