### 2018-07-05 edf8897
Adding documentation for the default sub-graphs view in the GCylc graph view.
