+++ doc/src/cylc-user-guide/gcylcrc.tex
@@ -115,6 +115,16 @@ be sorted using ascending or descending order.
     \item {\em example:} \lstinline@sort column ascending = False@
 \end{myitemize}
 
+\subsubsection{sub-graphs on}
+
+Set the sub-graphs view to be enabled by default.
+This can be changed later using the toggle options for the graph view.
+
+\begin{myitemize}
+\item {\em type:} boolean (False or True)
+\item {\em default:} ``False''
+\end{myitemize}
+
 
 \subsubsection{task filter highlight color}
 
