### 2018-07-06 44135d1
fix poll timers list override issue
