### 2018-07-06 48917e6
fix poll timers list override issue
