### 2018-07-10 c22bab7
improve poll logging
