### 2018-07-10 a5003f8
pep8 and grep fix
