### 2018-07-13 15ffc83
Fix restart-reload release runahead bug
For a suite with no final cycle point and tasks in the runahead point
(e.g. due to max active cycle points), a restart followed by a reload
would bring down the suite because the stop point was loaded from the
database as a string 'None' instead of a proper None.

This fixes the issue.
