### 2018-07-14 8d9579a
Avoid creating an unnecessary SuiteProcPool
