### 2018-07-16 75cced9
Improve logging
Basic stream logging for utility commands.
Single log file for suite server program.
Return control back to Python's logging module, where possible.
