+++ lib/cylc/suite_logging.py
@@ -1,646 +1,0 @@
-#!/usr/bin/env python2
-
-# THIS FILE IS PART OF THE CYLC SUITE ENGINE.
-# Copyright (C) 2008-2018 NIWA & British Crown (Met Office) & Contributors.
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <http://www.gnu.org/licenses/>.
-"""
-Provides loggers for use by cylc suites.
-"""
-from __future__ import print_function
-
-import os
-import sys
-import glob
-import logging
-import logging.handlers
-from time import time
-
-import cylc.flags
-from cylc.cfgspec.glbl_cfg import glbl_cfg
-from cylc.wallclock import (
-    get_time_string_from_unix_time, get_current_time_string)
-
-# Suite log filenames.
-SUITE_LOG = 'log'
-SUITE_OUT = 'out'
-SUITE_ERR = 'err'
-
-SUITE_LOG_OPTS = {
-    'l': SUITE_LOG,
-    'o': SUITE_OUT,
-    'e': SUITE_ERR,
-}
-
-LOG_DELIMITER = '.'
-
-
-def get_logs(directory, basename, absolute_path=True):
-    """Returns a list of log files in the given directory for the provided
-    basename (i.e. log, err, out) ordered newest to oldest."""
-    log_files = glob.glob(os.path.join(directory,
-                                       basename + LOG_DELIMITER + '*'))
-
-    if LOG_DELIMITER != '.':
-        old_logs = glob.glob(os.path.join(directory,
-                                          basename + '.*'))
-        new_logs = log_files
-    else:
-        old_logs = [log for log in log_files if len(log.split('.')[-1]) < 3]
-        new_logs = [log for log in log_files if log not in old_logs]
-    old_logs.sort()
-    new_logs.sort(reverse=True)
-    if absolute_path:
-        return new_logs + old_logs
-    else:
-        return [os.path.basename(log) for log in new_logs + old_logs]
-
-
-class StreamRedirectRoller(object):
-    """Redirect a stream to a rolling file.
-
-    Note that simply reassigning the sys streams is not sufficient
-    if we import modules that write to stdin and stdout from C
-    code - evidently the subprocess module is in this category!
-
-    Args:
-        stream (file): The stream to redirect.
-        link_path (str): The path to the symlink pointing at the log file.
-
-    """
-
-    def __init__(self, stream, link_path):
-        self.stream = stream
-        self.path = link_path
-
-    def doRollover(self):
-        file_handle = file(self.path, 'a+', 0)  # 0 => unbuffered
-        os.dup2(file_handle.fileno(), self.stream.fileno())
-
-
-class RollingFileHandler(logging.handlers.BaseRotatingHandler):
-    """A file handler for log files rotated by symlinking with support for
-       synchronised rotating of multiple logs."""
-
-    def __init__(self, filename, mode='a', maxBytes=0, encoding=None,
-                 file_stamp_fcn=None, archive_length=None):
-        logging.handlers.BaseRotatingHandler.__init__(
-            self, filename, mode, encoding, 0)
-        self.maxBytes = maxBytes
-        self.archive_length = archive_length
-        self.syncronised_group = None
[…]
