### 2018-07-17 c9d468a
Updated source file copyright statements.
