### 2018-07-17 a3b3f61
Add CONTRIBUTING.md and tidy README.md.
