+++ CONTRIBUTING.md
@@ -1,0 +1,103 @@
+# Cylc: How to Contribute
+
+## Report Bugs
+
+Report bugs by opening an issue on the [Cylc repository @
+Github](https://github.com/cylc/cylc/issues). Give your Cylc version number and
+a recipe to reproduce the bug.
+
+## Request Enhancements
+
+Request enhancements by opening an issue on the [Cylc repository @
+Github](https://github.com/cylc/cylc/issues). Describe the use case in detail.
+
+## Contribute Code
+
+All contributions to Cylc are made via pull requests against the *master*
+branch of [cylc/cylc](https://github.com/cylc/cylc). New contributors
+should add their details to the [Code Contributors](#code-contributors)
+section of this file as part of their first request. The developer who
+reviews each pull request is responsible for checking that the contributor's
+name is listed in this file before merging the pull request into *master*.
+
+## Code Contributors
+
+The following people have contributed to this code under the terms of
+the Contributor Licence Agreement and Certificate of Origin detailed
+below. (NOTE: those in _italics_ predate the introduction of this agreement;
+they will be contacted for consent).
+
+ -  Hilary Oliver
+ -  Matt Shin
+ -  Ben Fitzpatrick
+ -  Andrew Clark
+ -  Oliver Sanders
+ -  Declan Valters
+ -  Sadie Bartholomew
+ -  _Luis Kornblueh_
+ -  Kerry Day
+ -  Prasanna Challuri
+ -  David Matthews
+ -  _Tim Whitcomb_
+ -  _Scott Wales_
+ -  Tomek Trzeciak
+ -  _Thomas Coleman_
+ -  Bruno Kinoshita
+ -  _Annette Osprey_
+ -  _Jonathan Thomas_
+ -  Rosalyn Hatcher
+ -  _Domingo Manubens Gil_
+ -  Jonny Williams
+ -  _Milton Woods_
+ -  _Alex Reinecke_
+ -  _Chan Wilson_
+ -  _Kevin Pulo_
+ -  Lois Hugget
+ -  _Martin Dix_
+ -  _Ivor Blockley_
+
+(All contributors are identifiable with email addresses in the git version
+control logs or otherwise.)
+
+## Contributor Licence Agreement and Certificate of Origin
+
+By making a contribution to this project, I certify that:
+
+(a) The contribution was created in whole or in part by me and I have
+    the right to submit it, either on my behalf or on behalf of my
+    employer, under the terms and conditions as described by this file;
+    or
+
+(b) The contribution is based upon previous work that, to the best of
+    my knowledge, is covered under an appropriate licence and I have
+    the right or permission from the copyright owner under that licence
+    to submit that work with modifications, whether created in whole or
+    in part by me, under the terms and conditions as described by
+    this file; or
+
+(c) The contribution was provided directly to me by some other person
+    who certified (a) or (b) and I have not modified it.
+
+(d) I understand and agree that this project and the contribution
+    are public and that a record of the contribution (including my
+    name and email address) is retained for the full term of
+    the copyright and may be redistributed consistent with this project
+    or the licence(s) involved.
+
+(e) I, or my employer, grant to NIWA and all recipients of
+    this software a perpetual, worldwide, non-exclusive, no-charge,
+    royalty-free, irrevocable copyright licence to reproduce, modify,
+    prepare derivative works of, publicly display, publicly perform,
+    sub-licence, and distribute this contribution and such modifications
+    and derivative works consistent with this project or the licence(s)
+    involved or other appropriate open source licence(s) specified by
+    the project and approved by the
+    [Open Source Initiative (OSI)](http://www.opensource.org/).
+
+(f) If I become aware of anything that would make any of the above
[…]
