+++ README.md
@@ -1,11 +1,18 @@
 # The Cylc Workflow Engine
 
 [![Build Status](https://travis-ci.org/cylc/cylc.svg?branch=master)](https://travis-ci.org/cylc/cylc)
+[![Codacy Badge](https://api.codacy.com/project/badge/Grade/1d6a97bf05114066ae30b63dcb0cdcf9)](https://www.codacy.com/app/Cylc/cylc?utm_source=github.com&amp;utm_medium=referral&amp;utm_content=cylc/cylc&amp;utm_campaign=Badge_Grade)
 [![DOI](https://zenodo.org/badge/1836229.svg)](https://zenodo.org/badge/latestdoi/1836229)
 
 Cylc (“silk”) orchestrates complex distributed suites of interdependent cycling
-tasks. It was originally designed for environmental forecasting systems at
-[NIWA](https://www.niwa.co.nz).
+(or non-cycling) tasks. It was originally designed for environmental
+forecasting systems at [NIWA](https://www.niwa.co.nz) (but it is not
+specialized to environmental forecasting in any way).
+
+[Quick Installation](INSTALL.md) |
+[Web Site](https://cylc.github.io/cylc) |
+[User Guide and other documentation](https://cylc.github.io/cylc/documentation) |
+[How to Contribute](CONTRIBUTING.md)
 
 ### Copyright and Terms of Use
 
@@ -20,59 +27,11 @@ WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
 PARTICULAR PURPOSE.  See the GNU General Public License for more details.
  
 You should have received a copy of the GNU General Public License along with
-cylc.  If not, see [GNU licenses](http://www.gnu.org/licenses/).
+Cylc.  If not, see [GNU licenses](http://www.gnu.org/licenses/).
 
 ## Cylc Documentation
  * See [The Cylc Web Site](https://cylc.github.io/cylc)
 
-## Code Contributors
-<!--  git shortlog -s -n -->
- *  Hilary Oliver
- *  Matt Shin
- *  Ben Fitzpatrick
- *  Andrew Clark
- *  Oliver Sanders
- *  Declan Valters
- *  Sadie Bartholomew
- *  Luis Kornblueh
- *  Kerry Day
- *  Prasanna Challuri
- *  David Matthews
- *  Tim Whitcomb
- *  Scott Wales
- *  Tomek Trzeciak
- *  Thomas Coleman
- *  Bruno Kinoshita
- *  Annette Osprey
- *  Jonathan Thomas
- *  Rosalyn Hatcher
- *  Domingo Manubens Gil
- *  Jonny Williams
- *  Milton Woods
- *  Alex Reinecke
- *  Chan Wilson
- *  Kevin Pulo
- *  Lois Hugget
- *  Martin Dix
- *  Ivor Blockley
- 
-## Acknowledgement For Non-Cylc Work:
-Licences for non-cylc work included in this distribution can be found in the
-`licences/` directory.
- * `lib/cherrypy/`:
-    External software library released under a BSD license.
-    Minor modification to ignore an import warning.
-    See [cherrypy](http://www.cherrypy.org/).
- * `lib/isodatetime/`:
-    Unmodified external software library released under the LGPL license.
-    See [metomi/isodatetime](https://github.com/metomi/isodatetime/).
- * `lib/jinja2/`:
-    External software library released under a BSD license.
-    See [Jinja2](http://jinja.pocoo.org/).
- * `lib/markupsafe/`:
-    External software library released under a BSD license, used by Jinja2.
-    See [MarkupSafe](http://www.pocoo.org/projects/markupsafe/).
- * `lib/xdot.py`:
[…]
