### 2018-07-17 d0b6f26
Tweak README.md etc.
