### 2018-07-17 #737 @hjoliver
I've just posted Pull Request to Cylc, with a new contributor license agreement. We should be good to go when that's reviewed and merged to master ... I'll report back to confirm shortly.
