### 2018-07-17 8927d15
Make '--help' output accurate & consistent with rest of CLI
