### 2018-07-18 7182b8c
Tidy copyright message a bit.
