+++ doc/src/cylc-user-guide/cug-html.tex
@@ -28,7 +28,7 @@
 \renewcommand{\headsep}{10pt}
 \lhead{\leftmark}
 \rhead{\rightmark}
-\lfoot{Copyright (C) 2008-2018 NIWA & British Crown (Met Office) & contributors}
+\lfoot{Copyright (C) 2008-2018 NIWA & British Crown (Met Office) & Contributors.}
 \rfoot{\thepage}
 
 \usepackage{titlepic}  % off CTAN, held locally in cylc doc dir.
@@ -214,7 +214,7 @@ basicstyle=\color{basic}\ttfamily,
 User Guide \\
 \protect \input{cylc-version.txt} \\
 GNU GPL v3.0 Software License \\
-Copyright (C) 2008-2018 NIWA & British Crown (Met Office) & contributors}
+Copyright (C) 2008-2018 NIWA & British Crown (Met Office) & Contributors.}
 
 \author{Hilary Oliver}
 
