### 2018-07-18 #737 @hjoliver
@arfon - new `CONTRIBUTING.md` CLA just merged to Cylc master: https://github.com/cylc/cylc/blob/master/CONTRIBUTING.md
