### 2018-07-19 4ecd2f3
Update copyright notice in line with PR #2730
