+++ lib/cylc/cylc-review/static/js/search.js
@@ -1,6 +1,6 @@
 /******************************************************************************
  * THIS FILE IS PART OF THE CYLC SUITE ENGINE.
- * Copyright (C) 2008-2018 NIWA
+ * Copyright (C) 2008-2018 NIWA & British Crown (Met Office) & Contributors.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
