### 2018-07-23 #737 @hjoliver
@arfon - the Zenodo DOI for cylc-7.7.0 (as reviewed) is: [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1251084.svg)](https://doi.org/10.5281/zenodo.1251084)
